matchingVariograms

Function to create a permutation of spatial dataset X by starting with a randomly shuffled version of X and then smoothing it until the variogram of this permutation matches the variogram of the unmodified X

Usage

matchingVariograms(
  X.randomized,
  long,
  lat,
  delta,
  target_variog,
  prctile,
  ids,
  i,
  seed = 0
)

Arguments

X.randomized

numeric vector random permutation of the values of X across its x,y-coordinates.

long

numeric vector: y-coordinates of points in dataset X

lat

numeric vector: x-coordinates of points in dataset X

delta

numeric vector Given a point in dataset X, the percentage of neighbors which should be within the smoothing kernel. This can be a list or a single numeric. Input to `locfit::lp` as `nn`

target_variog

variogram variogram of dataset X that will be used as target when doing the matching. See geoR::variog

prctile

numeric: the max distance that should be used for evaluating the fit between the variograms from the permuted data and the target variogram. Can be determined from calculating all distances between the points and then calculating a specified quantile. Input to `geoR::variog` as `max.dist`

ids

integer: indices from 1 to the number of points in the dataset X

i

integer the number indicating the ith permutation

seed

integer Seed for the random number generator. Default 0.

Value

The output is returned as a list.

residusnumeric vector of length of delta, sum of squares of the residuals between the target variogram and the permuation variogram for each delta

delta.star.idinteger, the index of the delta that minimizes the residual sum of squares

hat.X.delta.starnumeric vector of length of X.randomized, permutation generated from optimal delta, whose variogram best matches the variogram of X

References

adapted from: Viladomat, Júlia et al. “Assessing the significance of global and local correlations under spatial autocorrelation: a nonparametric approach.” Biometrics vol. 70,2 (2014): 409-18. doi:10.1111/biom.12139

Examples

data(quakes)

#remove duplicated positions
quakes_data <- quakes[!duplicated(cbind(quakes$lat, quakes$long)),]

lat <- quakes_data$lat
long <- quakes_data$long
X <- quakes_data$depth
ids <- 1:length(X)

#number of permutations
B <- 3

#sequence of deltas
delta <- seq(0.1,0.9,0.1)

# maximum distance for the variogram set at the 25% percentile of
# the distribution of pairs of distances:
dists <- dist(cbind(lat, long))
prctile <- quantile(dists, probs = 0.25)

# variogram of variable X that will be used as target when doing the matching:
target_variog <- geoR::variog(data = X, coords = cbind(long, lat),
                              max.dist = prctile, option = "bin",
                              messages = FALSE)

# create B random permutations of the values of X across locations:
X.randomized <- lapply(1:B, function(i) {
   sample(X, size = length(X), replace = FALSE)
})

# create B permutations of the values of X with variograms that match the
#target variogram of X:
output <- lapply(1:B, function(i) {
  output <- matchingVariograms(X.randomized[[i]], long, lat,
                               delta, target_variog, prctile, ids, i)
})

output
#> [[1]]
#> [[1]]$residus
#> [1] 2022894773   22098926    5766723   17073605   14562351   14835429   10819006
#> [8]   38362235   40676504
#> 
#> [[1]]$delta.star.id
#> [1] 3
#> 
#> [[1]]$hat.X.delta.star
#>   [1] 5056.290 5019.708 4402.192 4993.113 5104.286 4735.094 4781.762 4329.019
#>   [9] 4399.851 5066.497 4898.679 4798.663 5069.868 5002.662 4851.001 4653.584
#>  [17] 4810.793 5087.615 4683.351 4781.920 5144.300 4815.049 4505.490 4407.441
#>  [25] 5121.019 5068.729 4803.808 5149.025 5079.951 5137.903 4654.666 4722.572
#>  [33] 4755.580 4530.943 4563.720 4550.393 4907.332 5009.064 5014.702 4827.018
#>  [41] 4720.580 4593.017 4496.610 4705.756 4873.805 4396.392 4337.316 4890.586
#>  [49] 4528.382 4565.888 5115.710 4706.558 4930.419 4966.227 5119.111 5204.463
#>  [57] 4702.011 4901.942 5027.738 4564.390 4999.185 5104.101 4835.108 4701.332
#>  [65] 4983.403 4763.569 5079.695 4703.807 5061.465 4652.838 4621.490 4742.792
#>  [73] 4745.934 4675.390 5070.847 4652.538 4829.649 4839.110 4916.405 4378.159
#>  [81] 4566.007 4568.651 5152.402 4965.872 5059.484 4750.855 4928.897 4608.417
#>  [89] 4686.105 4676.139 4730.717 4745.734 5053.488 4893.933 4732.787 5107.705
#>  [97] 4432.453 4851.136 4790.089 4404.071 4735.424 4476.263 4817.033 4600.282
#> [105] 4989.366 4813.243 4573.138 4828.053 4676.152 4677.104 4799.558 4999.421
#> [113] 5111.383 4484.326 4473.503 4473.871 4873.590 4734.285 4768.075 4949.618
#> [121] 4837.980 4942.563 4837.809 5057.296 4535.220 4793.506 4876.358 4511.955
#> [129] 4448.765 4851.561 4684.241 5004.388 4887.231 4594.684 4949.147 4865.081
#> [137] 4668.758 4475.378 4319.989 4799.209 4693.376 5109.916 4798.893 4419.237
#> [145] 4758.428 4712.784 4734.120 4904.371 4552.061 5104.404 4733.912 4734.254
#> [153] 4589.143 4791.268 4806.763 5017.009 4806.032 4517.978 5016.388 4743.899
#> [161] 5099.666 4499.579 4756.077 4549.818 4603.717 4635.536 4336.020 4784.294
#> [169] 4646.189 4937.397 5080.591 4459.245 4992.585 4704.385 4960.751 4574.659
#> [177] 4675.271 4711.259 5021.012 4876.215 5065.140 4424.770 4775.129 4978.164
#> [185] 5058.501 4646.953 4912.196 5048.544 4682.775 4761.234 4697.657 4782.871
#> [193] 5032.014 4559.178 4553.245 5105.728 4515.657 4642.151 5056.663 5079.465
#> [201] 5233.490 5085.672 4841.501 4328.016 4747.789 4916.251 4915.693 5111.938
#> [209] 4498.380 5066.775 4546.778 4854.611 4368.322 4604.025 4577.776 4983.026
#> [217] 4236.903 5035.288 4874.821 5010.505 4718.120 4915.729 4715.851 5089.557
#> [225] 4612.329 4763.417 5076.543 4742.433 4612.079 4929.359 5224.851 4495.564
#> [233] 4914.717 4470.599 4590.598 4624.165 4265.523 4893.467 4752.709 4529.650
#> [241] 5046.567 4511.842 4835.139 4812.451 4665.579 5076.815 5015.857 4675.900
#> [249] 4600.821 4876.154 4803.124 4745.537 5089.438 4972.253 4363.213 5023.853
#> [257] 4706.304 4903.136 4569.816 4470.861 4837.277 5143.973 4754.861 4986.410
#> [265] 4931.373 4423.621 4846.707 4782.233 5065.961 5020.799 5122.565 4835.777
#> [273] 4605.209 5177.081 4795.693 5149.356 4591.630 4443.586 4803.796 5028.226
#> [281] 4464.892 4774.894 4953.609 4683.921 4881.346 4571.007 5127.631 4774.405
#> [289] 5117.190 5145.060 5085.179 4805.323 5101.721 4450.740 4653.104 4689.710
#> [297] 4542.447 5070.980 4538.842 4963.172 4737.975 4764.701 4649.514 5030.737
#> [305] 4803.829 4688.987 5060.586 4865.278 5005.665 4415.194 4813.490 4788.569
#> [313] 5044.850 4908.800 4847.801 5002.553 5117.623 4909.096 4523.486 4796.893
#> [321] 4800.366 4850.428 5082.238 4468.304 4810.631 4327.891 5008.924 4772.900
#> [329] 4701.034 4758.571 4761.079 4595.584 4333.565 4835.923 4591.833 4349.663
#> [337] 4367.700 4288.833 4259.860 4249.743 4321.535 4313.483 4296.210 4289.555
#> [345] 4425.474 4407.532 4284.120 4429.213 4610.500 4379.890 4372.206 4751.029
#> [353] 5001.193 4468.642 4369.443 5009.784 4950.830 5008.363 4269.003 4664.950
#> [361] 5050.024 4656.329 4778.299 4603.421 4871.634 4358.928 5090.735 5009.276
#> [369] 4972.980 4708.322 4524.488 4803.908 5105.287 5109.462 4936.746 4686.467
#> [377] 5014.349 4706.651 4672.562 4641.860 4835.120 4813.120 4405.695 4812.135
#> [385] 5089.470 4419.743 5145.320 4373.557 4778.315 4365.785 4612.707 4305.417
#> [393] 5070.286 5110.276 4654.436 5081.691 4553.910 5053.949 4712.793 4441.072
#> [401] 4717.987 4507.585 4530.496 4739.473 5131.873 5069.110 4762.183 4617.342
#> [409] 4545.515 4563.829 4710.352 4944.714 4464.199 4772.658 4855.835 4824.436
#> [417] 4477.066 4473.239 4589.202 4564.698 4335.095 4672.588 4850.374 4537.984
#> [425] 4422.005 5024.239 5134.631 4843.963 5099.299 4882.943 4713.001 5050.671
#> [433] 4505.527 4892.381 4251.684 4875.948 5093.459 5184.342 4624.229 4754.671
#> [441] 4603.463 4926.977 5100.983 4814.822 4917.840 4726.356 5092.247 5015.023
#> [449] 4538.466 4254.704 4571.373 4923.351 4776.650 4594.013 4806.317 5106.672
#> [457] 4558.811 4633.093 5131.332 4674.872 5089.539 5038.149 5103.269 4380.328
#> [465] 5024.321 4868.960 4735.886 5069.452 4545.211 4724.193 4636.783 4733.014
#> [473] 4663.091 4541.237 4498.606 4593.408 4371.348 4590.184 4668.997 4875.703
#> [481] 4714.345 4871.106 4476.898 4977.077 4751.315 4555.543 5047.544 4966.647
#> [489] 4767.309 4491.185 4845.652 4687.007 4837.215 4754.644 4834.036 4770.590
#> [497] 4394.578 4606.273 4754.532 4761.054 4638.205 4819.162 4787.938 4619.561
#> [505] 5065.417 4764.575 4845.755 4763.669 4461.815 4412.693 4797.905 4594.968
#> [513] 4628.875 5047.350 4721.690 4686.071 4774.808 4644.528 4453.925 4666.666
#> [521] 5044.320 4518.867 4778.334 4598.200 4982.416 4844.572 4726.292 4857.144
#> [529] 4592.802 4849.831 4805.938 4997.245 4863.270 4657.574 4788.806 5140.876
#> [537] 4740.845 4783.439 5032.041 4803.594 4816.184 4884.976 4766.358 4883.049
#> [545] 4979.217 4812.767 5081.540 4578.506 5064.792 5112.306 4921.339 4811.954
#> [553] 4678.509 4652.720 4716.955 4685.039 4770.506 4847.783 4741.619 5027.771
#> [561] 4617.823 4461.068 4912.027 4452.520 4800.026 4224.746 5108.734 4844.694
#> [569] 4598.766 4823.348 4563.790 4970.650 4596.778 5087.483 4541.268 5154.808
#> [577] 5074.746 5040.505 4316.109 4716.716 5049.279 4865.832 5179.898 5104.048
#> [585] 4724.896 4879.233 4519.272 5009.687 4658.994 5072.130 5041.215 4785.615
#> [593] 4712.828 4764.957 4959.028 4912.734 4892.571 4600.960 4783.775 4325.729
#> [601] 4427.474 4886.916 4935.353 4839.574 4541.371 4726.927 4685.111 4571.291
#> [609] 4491.869 4294.892 4939.747 4833.127 4990.143 4606.705 5082.869 4804.825
#> [617] 4904.343 5151.244 4801.329 4551.882 4599.688 4618.094 4581.075 4951.523
#> [625] 4633.026 4534.632 4456.033 4831.410 5027.507 4906.287 4323.306 4425.247
#> [633] 4867.136 4885.084 5103.068 4711.388 4742.826 4665.532 4627.960 5038.663
#> [641] 4821.849 4284.435 4684.568 5004.433 4709.184 4636.396 4667.751 4635.856
#> [649] 5043.131 4990.678 5054.269 4814.215 4711.665 4858.818 4932.664 5066.887
#> [657] 4341.767 5083.744 5032.266 4367.962 5092.006 5036.225 4654.254 5135.569
#> [665] 5209.560 5130.968 4744.748 4585.882 4721.321 5059.710 4766.346 5042.060
#> [673] 5050.246 4845.846 4458.345 4695.731 4559.320 4688.457 5045.428 4840.514
#> [681] 4745.178 4489.365 5036.036 4626.536 4852.179 5201.390 4918.526 5093.897
#> [689] 5062.416 5052.464 5095.628 5144.613 5125.003 4515.942 4586.938 5052.446
#> [697] 4713.260 4850.963 5131.308 4855.282 5050.474 5058.622 4533.934 4651.058
#> [705] 4585.228 4625.502 4656.612 4984.957 5046.684 4725.593 4851.154 4606.312
#> [713] 4868.505 4833.302 4835.839 5005.539 4351.237 4610.978 5073.361 4897.985
#> [721] 5013.439 4534.646 4516.463 4809.957 4838.472 4991.025 4950.949 4564.102
#> [729] 5115.230 5072.797 4469.130 4352.746 4524.899 4777.947 4685.270 4902.659
#> [737] 4535.667 5146.072 4429.527 5143.766 4661.812 4693.795 4684.087 4995.002
#> [745] 4478.520 4458.242 4848.393 4814.056 4444.439 5075.569 4631.185 4865.191
#> [753] 4938.099 4605.585 4846.843 4750.292 4668.490 4700.051 4627.839 4835.832
#> [761] 4936.230 4670.498 4840.295 4842.836 4784.111 4698.337 4805.143 4801.617
#> [769] 4595.729 4652.664 4992.661 4754.467 4931.567 4906.269 4949.242 4879.955
#> [777] 5046.827 4817.503 4871.835 4834.916 4619.179 4640.198 4719.425 4307.778
#> [785] 4830.653 4651.754 4713.532 5055.376 5076.740 4566.936 5126.362 4357.709
#> [793] 5031.876 4580.096 5026.240 5088.328 4753.769 5001.382 4582.499 4660.407
#> [801] 4657.119 4912.954 4976.192 4610.898 5007.129 5046.711 5148.197 4416.808
#> [809] 4896.998 4650.281 4750.047 4687.307 5021.593 4785.047 4696.294 4592.803
#> [817] 4869.688 4790.229 4485.352 4578.032 4859.999 5060.528 4971.298 4997.272
#> [825] 4786.353 4722.604 4691.559 4837.840 4399.597 4799.693 4793.619 5128.776
#> [833] 5131.008 4631.020 4917.529 4564.363 4597.862 4917.692 5008.686 4647.426
#> [841] 4340.692 4889.879 4367.297 5088.706 4492.232 4428.285 4867.506 4962.001
#> [849] 4686.941 4834.120 4835.449 4616.159 5046.371 4634.869 4845.305 4731.252
#> [857] 4390.727 4840.315 4988.426 4895.976 5027.989 4959.363 4806.696 4971.295
#> [865] 4737.119 4700.518 4805.415 4732.252 5012.893 4755.918 4793.714 4861.755
#> [873] 5070.474 5064.081 5099.120 4870.319 4810.960 5107.363 4684.546 5001.800
#> [881] 4783.966 4778.330 4299.735 4568.289 4950.662 4808.310 4728.434 4691.114
#> [889] 4811.970 4778.138 4808.521 5047.022 5082.421 4595.765 4674.405 4911.462
#> [897] 4700.932 5088.134 4495.925 4580.234 4403.219 5144.624 5081.325 5091.214
#> [905] 4577.736 4921.118 4885.199 4355.200 4585.088 4807.497 4629.076 5036.751
#> [913] 4796.144 4830.170 4630.515 4702.309 4475.307 5054.478 4815.331 4801.831
#> [921] 5078.841 5134.995 4286.154 5054.124 4761.336 4328.808 4324.743 4867.132
#> [929] 4668.864 4588.102 4567.954 4669.481 4717.437 4574.368 5041.706 4841.131
#> [937] 4919.003 5019.509 5058.403 4261.240 5089.855 4682.380 4633.944 4303.112
#> [945] 5086.882 4309.259 4600.126 4655.581 5061.736 4575.756 4724.283 4510.327
#> [953] 4628.886 4543.328 4378.185 4684.464 4982.163 4843.814 4563.998 4877.850
#> [961] 4784.431 4714.915 4894.571 5062.005 4638.912 4996.756 4766.944 4615.606
#> [969] 4388.521 4648.381 4641.012 4761.116 4718.881 4656.780 4858.132 4311.787
#> [977] 5052.360 4924.522 5070.036 4969.684 4310.757 4304.513 4535.241 4847.176
#> [985] 4861.213 4904.723 5025.332 4730.071 5099.271 5088.192 4898.679 5137.400
#> [993] 4606.196 4495.205 4801.998 4821.946 4588.211 4896.464
#> 
#> 
#> [[2]]
#> [[2]]$residus
#> [1] 546621042 234616129 256865338 212255627 238233222 227193150  89885663
#> [8]  60206055  35610782
#> 
#> [[2]]$delta.star.id
#> [1] 9
#> 
#> [[2]]$hat.X.delta.star
#>   [1] 114932.1 114847.5 114370.1 114983.9 114978.4 114862.2 109004.7 114107.3
#>   [9] 114056.4 114722.0 114804.6 109279.1 115156.4 114913.3 110664.5 114852.2
#>  [17] 109124.7 115124.6 114719.6 114819.2 114871.4 109131.8 114783.6 113828.6
#>  [25] 114859.3 114984.0 109511.2 114769.2 114856.9 115004.5 114841.6 109392.3
#>  [33] 114860.7 114760.9 114876.8 114679.6 110078.0 114990.7 115025.7 110060.5
#>  [41] 112841.9 114846.5 114827.7 114807.5 110191.3 113857.8 114404.5 109053.5
#>  [49] 114810.7 114910.1 115073.5 114992.1 112771.0 114892.9 114953.4 114888.7
#>  [57] 114756.8 114868.9 114969.6 114912.1 114924.2 114853.1 111124.5 109074.6
#>  [65] 114863.5 114949.4 114796.8 114842.9 114894.4 114892.8 114848.9 114847.2
#>  [73] 109087.9 114711.7 114945.0 114808.3 114863.0 109365.7 114768.2 114060.7
#>  [81] 113387.5 114734.0 114949.3 114929.3 115035.4 114859.8 111073.3 114753.8
#>  [89] 114830.9 114864.7 109639.9 109347.2 114979.1 110383.3 114836.1 114963.8
#>  [97] 114296.1 115017.8 109284.4 114779.2 114934.9 114646.0 114787.8 113495.8
#> [105] 114776.1 114953.9 113323.6 110563.2 114668.2 112945.2 114854.7 114929.3
#> [113] 115032.6 114471.4 114563.8 114720.8 111084.2 111092.4 109609.6 110264.7
#> [121] 109390.3 114832.1 114830.3 115062.2 113683.0 109752.3 114907.8 114535.9
#> [129] 114140.8 115021.6 114893.3 114773.2 110195.6 114782.8 114905.9 109706.2
#> [137] 114909.4 114602.9 114577.5 114972.6 109954.4 114913.1 109480.6 113847.9
#> [145] 111743.0 114885.7 114885.2 112678.1 114818.3 114954.2 114648.8 109609.4
#> [153] 114639.9 108882.0 109614.8 114765.4 109125.0 114815.3 110702.0 109043.5
#> [161] 114969.8 114779.6 109444.8 113275.1 113232.1 113330.6 114443.3 114862.1
#> [169] 114862.9 110197.5 115082.1 114639.6 114760.8 114749.2 114727.8 113571.8
#> [177] 114790.6 115022.4 115019.5 114767.3 115103.2 114580.4 114804.5 114921.2
#> [185] 114892.4 114904.2 114764.7 114986.4 114873.9 114957.7 114840.0 109957.2
#> [193] 115049.8 114827.8 113570.8 114938.0 114669.2 114545.9 114900.1 114858.1
#> [201] 114911.2 115096.5 114842.9 114244.0 109986.2 114856.4 114835.2 114898.0
#> [209] 113884.6 114955.8 113820.2 114901.5 114194.3 114716.9 114834.0 114968.8
#> [217] 114354.5 115012.5 114847.4 115036.6 114961.8 111812.1 114664.3 114887.4
#> [225] 114782.4 109217.5 114851.3 114843.4 114611.9 111788.0 114982.2 114703.1
#> [233] 114871.1 114814.0 114817.0 114794.5 114165.6 114780.3 109296.5 114684.5
#> [241] 114994.1 114522.9 109158.0 115053.4 114789.9 115049.2 115084.5 114838.6
#> [249] 114829.0 110219.0 109117.8 109386.6 115062.0 111314.4 113904.2 114891.0
#> [257] 114832.3 110154.0 114869.7 114774.7 114875.7 114896.1 110061.8 115027.0
#> [265] 115068.9 114362.8 109431.3 109914.3 114890.7 114744.5 114932.6 114747.8
#> [273] 114732.1 114912.8 114811.8 115001.5 114843.2 114785.7 114774.1 115029.3
#> [281] 113719.7 114851.0 114345.3 115022.4 114983.5 114905.4 114976.2 114694.8
#> [289] 115017.0 114926.3 114917.1 109196.7 114871.3 114703.2 114898.3 114812.1
#> [297] 114933.5 114994.3 113824.7 110040.4 110108.2 114848.5 115013.3 114798.3
#> [305] 110548.5 114867.8 115052.0 114742.0 114789.2 113526.6 110163.5 110156.6
#> [313] 115032.8 114887.2 114789.5 114978.0 114942.6 110119.9 114641.0 109362.5
#> [321] 109475.5 114808.8 114898.6 113997.5 109145.7 114382.8 114848.5 108786.4
#> [329] 114747.7 109064.9 114953.9 114830.1 114245.5 109157.7 114799.0 114286.6
#> [337] 114471.5 114372.7 114242.4 114246.9 114218.8 114240.0 114280.1 114371.3
#> [345] 114392.6 114194.3 114199.2 114215.4 114707.1 114171.7 114038.0 109116.8
#> [353] 114957.3 113791.6 114268.9 114988.7 111615.7 115080.4 114318.2 109240.3
#> [361] 115122.8 114914.6 114890.4 113515.4 110032.1 114421.0 114853.1 115021.2
#> [369] 114848.6 114704.9 114781.4 114831.1 114959.6 114919.4 114785.3 114870.2
#> [377] 115069.3 114792.3 114890.1 114890.4 109605.8 109440.5 114496.2 109570.5
#> [385] 114895.0 114045.2 115008.0 114368.4 108879.7 114211.4 114761.0 114217.1
#> [393] 114834.3 115037.0 114813.1 114977.5 114883.1 114952.1 110268.2 113842.9
#> [401] 109074.4 113889.7 114591.4 114899.1 115015.6 114822.6 109589.5 114780.1
#> [409] 113553.9 114855.6 114639.3 110353.1 114725.3 114664.1 109522.5 109430.9
#> [417] 113567.0 113464.5 114702.9 114772.8 114385.3 114817.1 114775.9 113550.7
#> [425] 113508.4 114810.9 114898.8 110292.6 115077.6 115050.0 114807.1 115105.7
#> [433] 114606.2 114703.2 114206.8 110169.3 114886.6 114906.0 114854.1 109209.7
#> [441] 114796.0 109756.5 114996.6 114667.4 114676.0 114849.9 114867.0 114829.2
#> [449] 114639.5 114378.5 114863.6 109609.2 114759.7 114725.2 109509.8 115089.3
#> [457] 114814.8 114834.8 114930.0 114849.4 114929.2 114950.5 114979.8 113843.4
#> [465] 114910.4 109521.6 114863.6 114919.3 113752.7 114934.1 114956.3 114887.5
#> [473] 108956.6 114807.1 113536.9 112791.5 113974.1 114744.8 114801.8 114823.9
#> [481] 114784.7 114715.7 113602.4 114958.7 114656.7 113274.0 115010.5 114813.9
#> [489] 110311.6 113861.8 110025.2 114792.1 114871.5 114919.1 112612.4 114866.9
#> [497] 113833.0 114749.8 114899.7 114894.6 114694.1 114829.6 109336.3 114850.3
#> [505] 114950.1 109007.7 108956.7 108933.9 114714.9 114347.4 114868.3 114831.0
#> [513] 114818.9 115091.7 114875.5 108932.5 114808.3 114646.2 114628.2 114748.4
#> [521] 115044.9 114504.9 109654.1 113568.0 114784.0 109535.1 109384.3 114841.6
#> [529] 113419.1 109475.9 109391.3 114867.2 109478.2 114510.1 109539.3 114895.4
#> [537] 109354.9 109550.8 115114.8 109510.3 109466.3 109463.4 109408.7 109354.4
#> [545] 109497.8 109065.7 114895.3 114688.4 114868.9 114965.9 110373.5 109460.4
#> [553] 114574.6 114663.4 114656.2 114635.7 114582.4 114780.5 109219.4 115065.0
#> [561] 114670.1 113758.6 114774.8 114701.1 114892.5 114208.2 114954.7 114929.1
#> [569] 113250.7 109481.8 114772.2 114981.4 114729.2 115142.3 114829.1 115039.1
#> [577] 114854.1 115079.0 114322.4 109526.7 114928.7 109074.1 114933.5 114839.0
#> [585] 114846.7 110166.6 114869.3 114902.1 114953.6 114785.1 114842.5 109393.9
#> [593] 109133.4 114687.7 110694.0 109339.7 114945.7 114707.7 114669.4 114192.0
#> [601] 114213.8 114824.6 114862.7 112907.0 113458.4 114672.8 114744.8 114804.4
#> [609] 113437.9 113900.1 111584.4 109792.2 114987.8 114761.0 115089.0 114664.6
#> [617] 110229.9 115076.6 108994.9 113486.9 113339.5 114929.8 114809.1 111786.6
#> [625] 114808.5 113418.5 114730.7 112540.5 115095.5 114861.8 114044.2 114146.4
#> [633] 114894.8 114960.3 115100.4 115007.6 109487.9 114890.5 114771.0 114875.6
#> [641] 109674.6 114172.7 114750.7 114833.4 114562.8 112636.0 114855.2 113024.7
#> [649] 114895.8 115081.6 115105.4 108920.7 114842.8 110187.4 111955.1 114998.5
#> [657] 114240.1 114889.6 114909.4 113829.9 115059.0 115044.1 114673.9 114887.2
#> [665] 114865.8 114937.2 114840.0 114776.9 114951.2 114963.6 108989.3 115110.4
#> [673] 114936.5 109550.1 114553.5 114890.4 114763.9 114828.6 114950.3 110124.1
#> [681] 114838.0 114622.0 114988.1 114864.8 110039.5 114998.6 115013.9 114821.4
#> [689] 115078.2 114964.9 114985.2 114883.3 114816.4 114728.3 114824.5 115006.1
#> [697] 114912.3 109468.2 115106.4 109397.7 114456.4 114935.3 114804.2 114823.3
#> [705] 114879.9 114882.8 114799.4 114755.0 114874.4 114813.2 109621.8 114694.2
#> [713] 111095.1 114906.6 111450.3 114766.2 114464.1 114899.0 114996.1 111741.3
#> [721] 115013.3 114478.5 114575.3 109725.4 110146.3 115154.8 114908.7 114858.9
#> [729] 114768.0 115099.2 114498.5 113821.1 114742.8 109421.6 114950.3 114763.8
#> [737] 114821.8 114988.1 114657.5 114996.1 114815.9 114697.4 112595.0 114945.9
#> [745] 113788.5 113868.3 114698.7 114821.2 114576.1 115076.9 114690.1 114822.7
#> [753] 110658.8 114764.8 109383.2 114913.9 114705.2 109446.5 114865.6 114826.3
#> [761] 114869.7 114927.1 114845.1 109257.4 109495.4 114956.6 114861.2 109108.4
#> [769] 114770.2 114839.0 115051.9 114870.5 114791.5 114847.0 114801.9 114627.9
#> [777] 114917.6 108796.7 111584.4 111752.7 114580.0 114712.5 114840.2 114162.2
#> [785] 109319.4 114869.2 114651.4 114792.9 115054.0 114699.0 114990.2 114127.6
#> [793] 114781.1 113575.8 110180.7 114933.6 114642.3 115011.0 114654.9 114809.6
#> [801] 114615.8 110048.3 114839.7 114784.1 114837.3 114921.3 114908.2 114804.5
#> [809] 114811.4 114681.9 109549.6 114684.3 115166.8 114795.8 114884.4 114888.1
#> [817] 114764.5 114835.5 114739.9 114815.7 114864.1 114811.2 109974.2 114759.1
#> [825] 109423.7 114832.0 114860.7 114824.7 113747.2 114837.9 114939.1 114975.5
#> [833] 115008.2 114827.9 109772.1 114929.8 114740.9 110250.7 115118.9 114850.0
#> [841] 114431.7 110921.9 114557.9 115005.9 114631.5 114682.1 114864.5 114819.4
#> [849] 114805.0 109231.9 109560.0 114819.1 115041.7 114923.6 110353.2 114857.3
#> [857] 113892.1 109335.5 114896.2 114829.2 115133.1 114906.3 109127.3 110261.7
#> [865] 114861.5 114874.4 113035.2 109268.7 114911.5 109319.9 109298.9 114774.8
#> [873] 115058.0 115031.7 114999.3 114823.0 114796.3 115054.1 114939.0 114765.8
#> [881] 109100.6 109175.2 114358.6 114904.9 110258.8 110936.2 114629.6 112788.2
#> [889] 109438.0 114788.6 109349.1 114806.1 115023.7 114711.6 114627.8 114801.2
#> [897] 114774.5 114923.7 113542.1 114847.2 113503.2 114941.6 115031.1 115027.8
#> [905] 114841.1 111167.2 110260.5 114064.2 114857.7 111765.6 114838.0 114884.3
#> [913] 109027.4 110928.2 114828.6 114650.8 114603.9 115059.8 109671.7 114880.9
#> [921] 114996.9 114977.4 114300.6 114756.5 109407.9 114029.0 114409.2 109189.2
#> [929] 114978.7 114826.7 114757.9 114826.6 114906.7 114855.5 115024.1 114736.6
#> [937] 114894.3 115053.4 114824.6 114209.8 115004.9 114586.2 114610.7 113809.7
#> [945] 115017.1 114209.8 114608.5 114602.6 115010.0 113108.6 114920.8 114844.0
#> [953] 114821.9 114825.3 114294.8 114771.8 114856.0 114788.5 114501.2 109557.1
#> [961] 109277.8 114887.9 114720.5 115019.1 114787.2 114751.4 109417.9 114486.7
#> [969] 114527.2 114453.4 114688.1 109953.3 114671.0 114949.4 114819.6 114112.9
#> [977] 115074.7 111512.0 114847.9 111525.4 114210.4 114151.2 114754.7 109356.7
#> [985] 111809.3 111328.1 114996.9 114895.4 114913.6 114666.1 114878.8 114930.6
#> [993] 114752.2 114552.0 109344.8 114791.4 114804.2 111124.0
#> 
#> 
#> [[3]]
#> [[3]]$residus
#> [1] 1712237957   88327926   31177321   37189590   29075426   28572303   27496091
#> [8]   39932545   48119142
#> 
#> [[3]]$delta.star.id
#> [1] 7
#> 
#> [[3]]$hat.X.delta.star
#>   [1] 16831.06 16795.30 16272.44 17029.67 16731.96 16854.23 16374.68 16180.33
#>   [9] 15938.66 17109.84 16576.72 16410.88 16967.67 16654.57 16393.64 17006.69
#>  [17] 16470.39 17079.40 16367.78 16373.38 16789.25 16456.78 16455.12 15975.31
#>  [25] 16891.75 17118.50 16442.84 17101.12 16777.17 16837.81 16412.20 16296.58
#>  [33] 16983.59 16295.61 16896.92 16131.86 16354.55 17065.76 17035.73 16295.42
#>  [41] 15548.91 16880.95 16217.85 16965.80 16345.24 15960.83 16192.00 16409.80
#>  [49] 16250.79 16838.49 16939.97 16831.83 16442.83 16632.35 16784.21 16870.87
#>  [57] 16518.43 16573.24 17034.92 16234.05 16728.45 16725.69 16520.27 16380.91
#>  [65] 16810.32 17045.19 16840.75 16515.01 16887.03 16963.00 16983.30 16842.92
#>  [73] 16451.45 16342.12 16961.92 16330.82 16774.03 16403.83 16631.55 16077.19
#>  [81] 15844.78 16371.08 16859.93 16840.86 17167.48 16735.50 16343.33 16323.31
#>  [89] 16891.23 16991.81 16384.13 16358.86 16709.68 16368.68 16867.22 16808.82
#>  [97] 16160.87 16976.28 16436.24 16183.99 16961.47 16058.79 16524.62 15869.43
#> [105] 16609.13 16860.97 15685.97 16353.02 16567.91 15505.69 16993.83 17024.48
#> [113] 16887.09 16098.86 16105.43 16178.30 16536.86 16469.96 16326.47 16325.21
#> [121] 16319.20 16723.50 16528.44 16819.49 15935.57 16263.27 16880.42 16027.19
#> [129] 16117.43 16869.39 16947.08 17029.92 16323.48 16304.72 16649.00 16286.95
#> [137] 16859.08 16307.00 16287.29 16956.26 16441.51 16738.83 16393.96 16003.75
#> [145] 16675.81 16692.65 16935.39 16443.67 16342.44 17068.94 16588.82 16328.19
#> [153] 16243.44 16413.56 16331.42 16738.01 16391.48 16258.31 16353.68 16411.57
#> [161] 17020.31 16190.82 16416.41 15694.75 15560.02 15653.68 16112.74 16841.90
#> [169] 16397.04 16397.70 16999.01 16178.37 16831.68 16980.22 17026.62 15841.79
#> [177] 16485.68 17009.13 17033.29 16901.74 17057.27 16206.99 16711.99 16728.12
#> [185] 16808.07 16882.53 16581.87 16815.08 16320.64 16964.69 16807.22 16408.17
#> [193] 17041.20 16292.79 15729.15 16807.35 16405.49 16389.80 16787.99 17051.48
#> [201] 16826.26 17042.08 16967.52 16306.33 16385.51 16638.83 16557.51 16844.84
#> [209] 15952.05 16818.40 15875.99 16969.29 16078.33 16700.28 16420.05 16684.48
#> [217] 16263.12 17066.20 16786.69 17060.66 16888.93 16364.23 16536.02 16791.02
#> [225] 16255.19 16368.55 16666.34 16855.77 16380.48 16389.75 16939.88 16159.52
#> [233] 16998.53 16817.73 16239.82 16348.88 16155.90 16536.92 16316.66 16235.74
#> [241] 16799.39 16382.07 16316.86 17027.22 16970.17 17077.60 17082.20 16846.28
#> [249] 16417.90 16396.42 16460.97 16279.42 17091.03 16298.33 16067.06 16838.80
#> [257] 16988.37 16430.64 16332.16 16350.31 16822.94 16758.46 16416.95 17056.42
#> [265] 16862.44 16245.38 16355.84 16241.87 16829.47 16729.28 16884.79 16598.87
#> [273] 16301.70 16828.27 16542.52 17033.52 16379.16 16264.32 16567.52 16908.20
#> [281] 15814.30 16553.98 17077.31 16944.57 16965.75 16297.98 16888.74 16556.17
#> [289] 16988.89 16879.08 16966.46 16379.66 16781.10 16206.01 17069.30 16914.52
#> [297] 16152.26 17062.43 15906.79 16467.92 16383.11 16742.79 16979.12 16652.53
#> [305] 16491.01 16387.12 17063.94 16517.07 16695.22 15916.98 16460.05 16357.39
#> [313] 17048.68 16664.63 16669.64 17039.37 17176.70 16405.25 16079.52 16441.42
#> [321] 16416.15 16637.96 16821.20 16134.96 16308.19 16267.04 16712.66 16445.91
#> [329] 16601.46 16370.74 16912.74 16297.36 16126.81 16418.14 16437.52 16091.60
#> [337] 16216.71 16118.79 16152.63 16113.15 16229.90 16138.38 16145.94 16184.26
#> [345] 16146.80 16134.22 16147.93 16178.99 16577.24 16177.85 16031.20 16379.09
#> [353] 17031.70 15986.74 16177.80 17144.35 16346.73 17034.81 16158.12 16411.84
#> [361] 17045.56 16962.23 16534.82 15787.97 16294.29 16116.75 16808.61 17043.49
#> [369] 17079.37 16420.17 16354.25 16798.81 17059.65 16818.26 16710.05 16862.94
#> [377] 16990.38 16955.20 16970.79 16990.90 16339.52 16374.50 16248.08 16349.43
#> [385] 16861.94 16018.54 17052.59 16175.48 16479.15 16103.91 16718.31 16157.42
#> [393] 16645.98 17157.56 16243.34 17078.26 16725.16 17056.97 16420.72 15975.50
#> [401] 16414.83 15986.61 16052.30 16985.79 16988.38 16761.52 16359.87 16324.26
#> [409] 15818.27 16872.68 16587.13 16397.38 16149.61 16628.96 16311.72 16387.26
#> [417] 15856.15 15790.60 16696.86 16329.29 16190.12 16910.20 16585.90 15729.98
#> [425] 15860.94 16740.19 16735.00 16356.83 17047.46 16999.31 16362.81 16952.59
#> [433] 16095.02 16619.30 16090.81 16320.81 16778.53 16803.65 17033.83 16386.07
#> [441] 16700.33 16336.12 16790.57 16513.05 16634.81 16941.23 16828.88 16645.65
#> [449] 16065.60 16170.79 16279.13 16311.78 16499.27 16764.67 16409.56 16988.67
#> [457] 16982.68 16402.23 17027.37 16988.58 17138.40 16988.51 17041.01 15958.81
#> [465] 16650.25 16374.29 17016.66 17043.10 15820.87 16867.54 17023.29 16934.18
#> [473] 16413.32 16705.70 15724.56 15426.75 16056.43 16360.39 16801.31 16631.90
#> [481] 16997.24 16490.70 15713.16 16984.42 16439.59 15760.79 17133.69 16754.62
#> [489] 16482.15 15889.13 16399.12 16303.16 16930.32 16926.91 16334.34 16907.46
#> [497] 15880.09 16295.01 16831.03 16834.45 16724.58 16631.15 16419.38 16315.02
#> [505] 16965.53 16368.84 16359.55 16427.14 16116.21 16134.83 16698.47 16832.56
#> [513] 16419.75 17135.19 16901.36 16464.96 16834.37 16768.30 16155.69 16699.34
#> [521] 16921.30 16205.13 16371.94 15808.55 16665.07 16375.57 16422.95 17121.34
#> [529] 15729.47 16371.46 16324.68 16774.22 16404.80 16502.24 16302.96 17063.38
#> [537] 16387.08 16374.66 17081.18 16398.28 16384.06 16446.07 16413.73 16430.51
#> [545] 16258.40 16476.83 16874.53 16847.42 16645.22 16811.19 16258.84 16293.16
#> [553] 16359.68 16401.51 16417.85 16466.84 16468.17 16539.48 16402.12 17014.49
#> [561] 16295.37 15887.24 16625.78 16133.92 16777.87 16042.86 17005.62 16758.36
#> [569] 15698.47 16379.92 16850.10 17023.37 16678.25 16984.84 16889.44 17060.28
#> [577] 16763.09 17047.99 16198.21 16344.97 16852.02 16405.03 16838.28 16750.40
#> [585] 16980.11 16280.50 16172.21 16657.05 16365.80 16788.47 16668.65 16415.47
#> [593] 16437.86 16559.33 16375.80 16506.21 17047.02 16343.23 16358.65 16097.76
#> [601] 16174.40 16593.51 16677.21 16592.88 15778.27 16615.03 16632.18 16266.44
#> [609] 15730.93 15958.17 16437.87 16332.19 16866.09 16954.95 17077.04 16523.92
#> [617] 16391.90 16991.40 16474.86 15779.95 15688.67 16257.12 16354.24 16487.45
#> [625] 16337.28 15846.07 16362.58 16418.72 16846.04 16562.30 16072.44 16084.04
#> [633] 16792.57 16743.24 16980.26 16888.90 16413.08 16843.81 16312.09 16647.00
#> [641] 16325.75 16121.63 16714.98 16748.26 16389.85 15537.19 16979.31 15519.72
#> [649] 16725.19 17100.09 17050.15 16288.48 16401.38 16410.53 16358.19 17117.32
#> [657] 16120.97 17055.43 17074.23 16030.49 16889.47 17018.38 16408.46 16974.76
#> [665] 16995.57 17008.44 16932.76 16336.78 16869.10 17139.72 16459.82 16990.11
#> [673] 16907.56 16390.78 16297.67 16954.16 16343.22 16406.62 16811.68 16285.57
#> [681] 16782.50 16248.86 16891.00 16959.49 16361.20 17006.04 17098.81 16738.07
#> [689] 17093.97 17096.76 16935.48 16774.91 16726.25 16139.48 16324.04 17034.88
#> [697] 17028.73 16293.59 16996.62 16408.35 17136.72 17062.90 16182.02 16761.77
#> [705] 16874.99 16397.50 16848.03 16636.26 16826.11 16961.74 16366.44 16639.10
#> [713] 16445.46 16522.48 16606.88 16712.57 16242.35 16831.98 17030.31 16377.65
#> [721] 17091.07 16002.40 16127.18 16316.20 16414.35 17029.49 16872.44 16380.06
#> [729] 16642.21 17044.25 16166.34 15825.54 16132.50 16389.23 16863.69 16544.10
#> [737] 16154.98 16742.26 16269.38 16651.22 16844.75 16603.70 15414.21 16958.74
#> [745] 15799.08 15865.95 16513.18 16603.88 16136.35 16981.00 16579.77 16751.05
#> [753] 16396.17 16175.03 16346.21 16917.09 16617.16 16417.45 16872.81 16712.46
#> [761] 16663.59 16943.34 16648.90 16370.75 16398.21 16739.30 16712.10 16406.63
#> [769] 16597.12 16675.93 17072.91 16727.51 16604.22 16531.71 16510.96 16635.87
#> [777] 16714.47 16446.70 16369.16 16334.91 16335.20 16722.33 16899.75 16197.62
#> [785] 16416.90 16349.85 16632.56 16780.71 17077.82 16157.93 16737.16 16165.60
#> [793] 16650.06 15800.63 16354.10 16824.22 16465.93 17083.89 16401.60 16318.73
#> [801] 16431.61 16406.42 16729.01 16802.80 16662.60 16827.23 16626.36 16210.44
#> [809] 16562.94 16845.75 16270.67 16417.96 17054.06 16575.16 17060.34 16928.28
#> [817] 16558.12 16476.85 16189.60 16723.06 16507.12 16734.29 16311.74 16630.76
#> [825] 16306.94 16949.52 16901.88 16667.38 15865.77 17026.71 16882.13 16844.08
#> [833] 16983.81 16954.52 16298.36 16262.58 16669.01 16326.91 17023.93 16917.26
#> [841] 16140.10 16302.00 16078.99 17015.74 16183.79 16110.55 16448.88 16644.80
#> [849] 16954.33 16476.33 16336.77 16819.38 17106.78 16831.53 16398.73 16984.36
#> [857] 15964.62 16517.76 16668.72 16724.77 16985.96 16656.12 16429.00 16353.90
#> [865] 16484.67 16969.81 16511.32 16433.19 16698.96 16427.89 16379.44 16515.57
#> [873] 17080.07 16919.35 16982.60 16540.50 16488.54 16998.28 16933.16 16676.43
#> [881] 16379.37 16457.68 16126.36 16260.72 16347.67 16321.22 16443.43 15491.27
#> [889] 16364.72 16553.92 16455.32 16603.30 17116.19 16285.25 16642.25 16482.85
#> [897] 16964.35 16983.96 15787.52 16770.68 15811.17 16805.44 17095.86 17139.64
#> [905] 16245.00 16394.83 16369.72 16074.81 16405.64 16407.98 16945.21 16693.77
#> [913] 16457.20 16289.01 16776.44 16448.07 16217.36 17009.12 16385.94 16690.80
#> [921] 17046.38 17015.97 16116.71 16681.22 16444.28 16050.17 16157.60 16386.11
#> [929] 16872.04 16338.65 16305.75 16731.75 16730.51 16776.44 16908.78 16543.84
#> [937] 16677.14 17003.72 16735.31 16060.92 16902.06 16398.41 16377.39 15993.94
#> [945] 16971.65 16157.98 16530.54 16394.72 16992.48 15595.92 16858.82 16294.63
#> [953] 16520.89 16296.95 16240.42 16645.32 16655.80 16651.07 16369.24 16401.96
#> [961] 16462.00 16875.84 16579.02 17012.99 16363.28 16661.15 16304.25 16369.61
#> [969] 16248.35 16322.79 16439.84 16329.06 16478.38 16872.30 16717.95 16081.79
#> [977] 17017.59 16307.90 16765.33 16378.23 16087.77 16174.15 16949.34 16402.78
#> [985] 16317.20 16438.56 17014.38 17072.12 16799.96 17009.40 16820.49 17095.61
#> [993] 16841.03 16107.72 16404.08 16813.80 16894.39 16339.29
#> 
#>