matchingVariograms

Function to create a permutation of spatial dataset X by starting with a randomly shuffled version of X and then smoothing it until the variogram of this permutation matches the variogram of the unmodified X

Usage

matchingVariograms(
  X.randomized,
  long,
  lat,
  delta,
  target_variog,
  prctile,
  ids,
  i,
  seed = 0
)

Arguments

X.randomized

numeric vector random permutation of the values of X across its x,y-coordinates.

long

numeric vector: y-coordinates of points in dataset X

lat

numeric vector: x-coordinates of points in dataset X

delta

numeric vector Given a point in dataset X, the percentage of neighbors which should be within the smoothing kernel. This can be a list or a single numeric. Input to `locfit::lp` as `nn`

target_variog

variogram variogram of dataset X that will be used as target when doing the matching. See geoR::variog

prctile

numeric: the max distance that should be used for evaluating the fit between the variograms from the permuted data and the target variogram. Can be determined from calculating all distances between the points and then calculating a specified quantile. Input to `geoR::variog` as `max.dist`

ids

integer: indices from 1 to the number of points in the dataset X

i

integer the number indicating the ith permutation

seed

integer Seed for the random number generator. Default 0.

Value

The output is returned as a list.

residusnumeric vector of length of delta, sum of squares of the residuals between the target variogram and the permuation variogram for each delta

delta.star.idinteger, the index of the delta that minimizes the residual sum of squares

hat.X.delta.starnumeric vector of length of X.randomized, permutation generated from optimal delta, whose variogram best matches the variogram of X

References

adapted from: Viladomat, Júlia et al. “Assessing the significance of global and local correlations under spatial autocorrelation: a nonparametric approach.” Biometrics vol. 70,2 (2014): 409-18. doi:10.1111/biom.12139

Examples

data(quakes)

#remove duplicated positions
quakes_data <- quakes[!duplicated(cbind(quakes$lat, quakes$long)),]

lat <- quakes_data$lat
long <- quakes_data$long
X <- quakes_data$depth
ids <- 1:length(X)

#number of permutations
B <- 3

#sequence of deltas
delta <- seq(0.1,0.9,0.1)

# maximum distance for the variogram set at the 25% percentile of
# the distribution of pairs of distances:
dists <- dist(cbind(lat, long))
prctile <- quantile(dists, probs = 0.25)

# variogram of variable X that will be used as target when doing the matching:
target_variog <- geoR::variog(data = X, coords = cbind(long, lat),
                              max.dist = prctile, option = "bin",
                              messages = FALSE)
#> Warning: no display name and no $DISPLAY environment variable

# create B random permutations of the values of X across locations:
X.randomized <- lapply(1:B, function(i) {
   sample(X, size = length(X), replace = FALSE)
})

# create B permutations of the values of X with variograms that match the
#target variogram of X:
output <- lapply(1:B, function(i) {
  output <- matchingVariograms(X.randomized[[i]], long, lat,
                               delta, target_variog, prctile, ids, i)
})

output
#> [[1]]
#> [[1]]$residus
#> [1] 986946227  77599218  11904997  24474093  23120332  13986652  21741995
#> [8]  24084522   3471635
#> 
#> [[1]]$delta.star.id
#> [1] 9
#> 
#> [[1]]$hat.X.delta.star
#>   [1] 26548.82 26571.45 26671.98 26243.09 26547.79 26256.29 27657.46 26729.82
#>   [9] 26761.65 26424.63 26668.10 27522.97 26408.14 26608.03 27562.47 25983.03
#>  [17] 27703.92 26395.65 26863.54 26846.58 26606.22 27722.48 26888.95 26689.60
#>  [25] 26582.78 26274.61 27563.06 26339.33 26571.20 26422.02 26942.35 27620.60
#>  [33] 25990.58 26862.45 25971.50 27094.15 27652.55 26297.91 26303.75 27594.33
#>  [41] 26751.68 25903.59 27113.75 25875.42 27605.15 26687.44 26833.81 27605.35
#>  [49] 26999.36 26212.29 26314.70 26321.46 27382.21 26749.16 26445.21 26346.59
#>  [57] 26744.01 26810.55 26271.07 27130.04 26588.50 26486.33 27136.27 27685.36
#>  [65] 26545.45 26062.94 26378.72 26863.83 26478.84 25843.01 25927.47 26269.65
#>  [73] 27704.03 26868.57 26259.38 26942.56 26191.51 27585.62 26576.47 26828.04
#>  [81] 26739.70 26863.25 26321.73 26381.26 26246.41 26250.82 27565.98 26901.93
#>  [89] 25879.26 25902.18 27691.37 27550.20 26574.12 27602.07 25953.49 26359.17
#>  [97] 26847.25 26152.82 27661.73 27014.42 26050.10 27139.46 26730.36 26656.81
#> [105] 26627.34 26297.44 26709.52 27698.01 26489.16 26728.78 25973.26 26106.98
#> [113] 26307.51 26957.21 27060.19 27205.47 27177.21 27119.42 27691.28 27731.28
#> [121] 27601.09 26790.96 26789.35 26366.73 26771.20 27728.50 26235.18 27075.03
#> [129] 26728.03 26324.15 26012.30 26171.40 27632.38 26923.23 26737.27 27659.78
#> [137] 25950.83 26690.40 26846.98 26232.97 27218.57 26591.51 27668.78 26711.81
#> [145] 27007.53 26344.78 26014.93 27408.41 26950.47 26237.64 26605.05 27639.94
#> [153] 26711.92 27660.82 27630.05 26456.54 27785.92 27077.41 27567.56 27653.94
#> [161] 26472.79 27002.43 27606.21 26745.30 26722.07 26761.61 26814.58 26283.09
#> [169] 26987.75 27694.48 26282.01 27012.27 26452.28 25912.25 26129.38 26787.01
#> [177] 26795.72 26134.40 26297.81 25971.49 26352.01 26756.80 26230.11 26399.11
#> [185] 26482.67 26072.73 26628.45 26485.93 27080.94 26079.40 26284.33 27560.46
#> [193] 26221.56 27029.57 26832.08 26605.90 26744.82 26643.73 26499.42 26377.67
#> [201] 26332.56 26198.15 26013.98 26650.15 27419.99 26630.61 26771.66 26491.30
#> [209] 26846.36 26443.28 26865.77 26058.22 26841.87 26243.59 26958.45 26663.83
#> [217] 26793.47 26322.84 26318.90 26198.40 26068.07 27525.74 26482.15 26431.18
#> [225] 26934.47 27555.23 26536.01 25944.77 26668.18 27592.43 26566.77 26934.16
#> [233] 26090.74 26213.83 26976.55 26933.96 26764.25 26677.75 27685.93 26965.11
#> [241] 26432.15 26678.67 27652.32 26201.16 25903.97 26325.41 26341.40 25851.77
#> [249] 26947.28 27633.91 27737.76 27517.68 26355.56 27539.69 26633.58 26597.86
#> [257] 25963.72 27597.55 27054.52 26939.18 26166.74 26595.15 27600.97 26312.83
#> [265] 26398.93 26677.34 27561.17 27571.78 26370.77 26573.82 26488.33 26714.14
#> [273] 26897.78 26439.30 26781.37 26307.44 26949.58 27055.25 26717.76 26305.34
#> [281] 26776.84 26827.09 26606.25 26130.27 26120.23 27124.45 26606.38 26315.29
#> [289] 26308.64 26481.69 26283.46 27709.59 26574.80 27039.42 25943.97 25914.47
#> [297] 27189.69 26280.56 26792.15 27620.89 27482.10 26240.53 26116.35 26540.58
#> [305] 27468.82 26996.02 26335.77 26685.24 26569.78 26667.30 27421.10 27382.03
#> [313] 26333.94 26658.61 26462.10 26372.64 26345.90 27682.23 27098.86 27560.09
#> [321] 27582.68 26531.33 26367.10 26757.94 27643.66 26761.46 26604.68 27659.38
#> [329] 26315.57 27671.13 26249.61 27000.86 26750.08 27568.32 26974.20 26790.83
#> [337] 26820.36 26876.68 26749.78 26742.73 26724.74 26786.04 26782.35 26901.46
#> [345] 26889.55 26747.00 26742.69 26734.85 26386.64 26671.65 26721.21 27668.19
#> [353] 26278.94 26648.42 26776.55 26294.95 27549.58 26181.43 26808.79 27729.52
#> [361] 26397.99 26028.57 26914.12 26749.70 27594.15 26819.20 26369.86 26173.62
#> [369] 26102.53 26703.72 26936.32 26255.32 26314.99 26478.27 26275.31 25881.65
#> [377] 26372.32 25813.11 25922.64 25906.29 27575.83 27612.46 26791.72 27673.49
#> [385] 26400.48 26765.28 26281.64 26922.09 27648.63 26622.65 26274.98 26654.37
#> [393] 26570.34 26230.42 26980.64 26252.89 25862.63 26269.77 27315.59 26760.30
#> [401] 27712.70 26813.44 27077.77 26057.24 26304.99 26362.36 27645.73 26913.23
#> [409] 26728.75 26017.18 26428.01 27620.77 27017.72 26339.01 27648.01 27557.29
#> [417] 26696.76 26690.33 26127.07 26960.86 26801.71 25937.48 26801.72 26738.14
#> [425] 26604.12 26502.65 26425.80 27702.02 26474.94 26222.97 26909.86 26336.74
#> [433] 27092.71 26430.28 26727.80 27655.11 26391.34 26487.19 25968.71 27694.51
#> [441] 26402.19 27628.92 26390.52 26510.59 26478.86 25946.39 26549.64 26559.71
#> [449] 26967.85 26791.06 27060.31 27645.42 26718.46 26296.41 27616.30 26374.28
#> [457] 26129.39 26970.01 26429.16 25879.27 26410.02 26423.63 26415.91 26723.37
#> [465] 26561.34 27583.44 25973.43 26259.53 26914.47 26244.97 26122.57 26002.46
#> [473] 27685.63 26384.40 26689.96 26682.65 26636.34 26873.43 26196.87 26682.40
#> [481] 25938.59 26587.87 26750.56 26284.23 26670.37 26632.39 26338.60 26582.73
#> [489] 27309.66 26751.47 27692.21 26908.81 26057.79 26034.96 27456.38 25981.81
#> [497] 26727.58 26926.33 26174.87 25973.09 26174.52 26435.22 27613.89 26973.85
#> [505] 26309.60 27705.73 27667.65 27707.27 27081.75 26721.75 26449.52 25947.89
#> [513] 26939.17 26292.11 25841.92 27623.53 26368.46 26053.92 27080.99 26302.06
#> [521] 26335.43 26698.65 27742.81 26723.80 26616.16 27676.17 27606.12 26115.95
#> [529] 26688.72 27655.51 27614.20 26461.46 27664.23 26457.72 27659.81 26193.22
#> [537] 27605.73 27651.30 26362.47 27593.04 27682.46 27678.73 27627.37 27560.78
#> [545] 27558.76 27691.64 26350.81 25696.84 26616.05 26536.60 27690.11 27756.99
#> [553] 26521.00 26632.85 26580.09 26640.09 26481.79 26735.14 27619.03 26344.57
#> [561] 26714.86 26669.22 26522.45 26979.53 26318.54 26773.97 26281.14 26303.37
#> [569] 26686.84 27607.24 26105.14 26115.93 26253.07 26352.10 25892.12 26340.04
#> [577] 26472.56 26251.89 26782.37 27668.97 26605.32 27633.60 26459.09 26569.66
#> [585] 25940.19 27529.12 27152.56 26618.98 27103.90 26547.60 26596.75 27561.73
#> [593] 27651.38 26405.54 27694.14 27657.00 26057.95 26794.22 26591.30 26796.97
#> [601] 26818.85 26651.15 26683.08 27140.70 26752.82 26418.95 26427.73 27015.48
#> [609] 26715.95 26721.42 27515.33 27589.04 26446.83 25822.39 26347.29 26420.35
#> [617] 27612.85 26373.46 27741.06 26693.84 26830.35 27150.86 26955.93 27594.34
#> [625] 26988.25 26650.70 27055.18 27467.68 26288.33 26691.53 26635.03 26728.79
#> [633] 26715.21 26403.59 26378.47 26368.34 27567.62 26028.78 26938.47 26668.49
#> [641] 27651.74 26723.76 26322.71 26521.31 26521.12 26698.84 25877.65 26838.02
#> [649] 26554.10 26315.83 26326.44 27606.93 26929.68 27346.37 27464.09 26377.28
#> [657] 26795.33 26368.51 26385.44 26746.56 26381.09 26487.92 26799.20 26336.35
#> [665] 26320.81 26375.08 25978.07 26936.63 26177.39 26271.58 27699.10 26378.97
#> [673] 26250.23 27712.36 26790.90 25971.12 26882.60 26973.17 26578.45 27624.51
#> [681] 26305.79 26956.60 26262.59 25902.06 27648.46 26465.93 26076.68 26522.13
#> [689] 26341.12 26277.43 26224.93 26419.99 26491.98 27050.08 26997.89 26295.80
#> [697] 26033.12 27575.90 26392.43 27552.93 26701.25 26227.99 27091.74 26185.27
#> [705] 25913.15 27036.50 25968.74 26442.12 26558.15 25834.42 27591.16 26363.21
#> [713] 27620.78 26868.45 27149.47 26511.74 26836.24 26111.42 26089.41 27597.99
#> [721] 26301.39 27026.85 27117.94 27668.47 27622.05 26255.30 26215.43 26990.45
#> [729] 26498.82 26394.27 26895.09 26774.21 27021.61 27553.52 25956.89 26527.27
#> [737] 27155.30 26617.87 26962.42 26643.01 26156.27 26458.92 26694.61 26258.92
#> [745] 26774.89 26812.40 26546.40 26769.12 26940.28 26307.47 26333.88 26625.68
#> [753] 27644.19 26983.33 27614.70 26032.49 26325.90 27511.35 25947.11 26377.55
#> [761] 26675.83 26057.87 26530.81 27562.28 27567.38 26263.87 26283.74 27574.91
#> [769] 26331.70 26326.21 26229.32 26255.19 26668.87 26747.63 26673.13 26354.30
#> [777] 26709.56 27653.43 27469.80 27599.82 26635.63 26232.33 26011.92 26684.78
#> [785] 27646.06 26999.27 26334.62 26525.89 26328.58 27161.46 26536.04 26734.08
#> [793] 26563.25 26727.47 27665.09 26489.19 26541.20 26296.06 26676.80 26924.02
#> [801] 26546.50 27316.94 26625.19 26106.63 26587.01 26469.30 26544.37 27099.99
#> [809] 26702.85 26063.00 27706.11 26729.90 26429.56 26754.93 26011.92 25962.52
#> [817] 26666.19 26799.80 27031.72 26193.76 26899.78 26493.62 27649.96 26469.69
#> [825] 27681.63 25946.92 25979.35 26546.99 26735.72 25963.95 26259.30 26359.40
#> [833] 26366.79 25914.21 27668.51 27139.42 26325.70 27722.81 26387.23 25967.58
#> [841] 26725.43 27576.28 26988.77 26333.89 26999.66 27033.02 26835.42 26622.50
#> [849] 25871.26 27598.21 27689.44 26125.34 26342.74 26160.49 27355.21 25979.40
#> [857] 26785.33 27475.47 26700.35 26677.94 26335.89 26577.54 27521.82 27662.21
#> [865] 26841.30 25993.34 27267.83 27504.99 26604.38 27555.16 27558.53 26723.08
#> [873] 26390.15 26290.19 26285.53 26685.09 26810.35 26337.03 26147.95 26521.19
#> [881] 27691.00 27737.52 26791.62 27140.94 27548.45 27615.86 26506.82 26738.70
#> [889] 27630.69 26778.57 27723.25 26413.67 26337.85 26875.07 26272.50 26761.32
#> [897] 25904.10 26245.60 26767.96 26208.47 26724.49 26524.07 26329.74 26348.71
#> [905] 27091.56 27643.25 27663.56 26659.60 26980.33 27580.05 25902.96 26496.07
#> [913] 27787.40 27651.07 26208.58 26583.22 26851.56 26347.44 27633.07 26410.47
#> [921] 26267.83 26276.72 26854.33 26529.26 27516.12 26745.40 26749.27 27719.87
#> [929] 26089.66 26945.91 26923.49 26408.31 26391.40 26225.64 26296.90 26554.93
#> [937] 26619.79 26293.99 26577.65 26792.50 26435.83 26592.81 26629.71 26665.35
#> [945] 26302.17 26758.74 26582.33 26619.00 26271.51 26751.59 26164.42 26975.33
#> [953] 26950.27 26980.95 26732.92 26353.65 26602.92 26498.21 26542.65 27633.98
#> [961] 27627.34 26282.09 26481.50 26295.23 26913.72 26605.68 27603.88 26535.10
#> [969] 26867.44 26475.01 26759.26 27603.32 26617.74 26065.34 26379.24 26654.36
#> [977] 26345.53 27473.40 26516.59 27554.63 26757.09 26680.97 25756.56 27625.91
#> [985] 27525.85 27530.70 26316.59 26042.48 26600.37 26221.73 26324.87 26248.60
#> [993] 25956.02 26978.57 27562.68 26254.69 25965.77 27570.07
#> 
#> 
#> [[2]]
#> [[2]]$residus
#> [1] 546621042 234616129 256865338 212255627 238233222 227193150  89885663
#> [8]  60206055  35610782
#> 
#> [[2]]$delta.star.id
#> [1] 9
#> 
#> [[2]]$hat.X.delta.star
#>   [1] 114932.1 114847.5 114370.1 114983.9 114978.4 114862.2 109004.7 114107.3
#>   [9] 114056.4 114722.0 114804.6 109279.1 115156.4 114913.3 110664.5 114852.2
#>  [17] 109124.7 115124.6 114719.6 114819.2 114871.4 109131.8 114783.6 113828.6
#>  [25] 114859.3 114984.0 109511.2 114769.2 114856.9 115004.5 114841.6 109392.3
#>  [33] 114860.7 114760.9 114876.8 114679.6 110078.0 114990.7 115025.7 110060.5
#>  [41] 112841.9 114846.5 114827.7 114807.5 110191.3 113857.8 114404.5 109053.5
#>  [49] 114810.7 114910.1 115073.5 114992.1 112771.0 114892.9 114953.4 114888.7
#>  [57] 114756.8 114868.9 114969.6 114912.1 114924.2 114853.1 111124.5 109074.6
#>  [65] 114863.5 114949.4 114796.8 114842.9 114894.4 114892.8 114848.9 114847.2
#>  [73] 109087.9 114711.7 114945.0 114808.3 114863.0 109365.7 114768.2 114060.7
#>  [81] 113387.5 114734.0 114949.3 114929.3 115035.4 114859.8 111073.3 114753.8
#>  [89] 114830.9 114864.7 109639.9 109347.2 114979.1 110383.3 114836.1 114963.8
#>  [97] 114296.1 115017.8 109284.4 114779.2 114934.9 114646.0 114787.8 113495.8
#> [105] 114776.1 114953.9 113323.6 110563.2 114668.2 112945.2 114854.7 114929.3
#> [113] 115032.6 114471.4 114563.8 114720.8 111084.2 111092.4 109609.6 110264.7
#> [121] 109390.3 114832.1 114830.3 115062.2 113683.0 109752.3 114907.8 114535.9
#> [129] 114140.8 115021.6 114893.3 114773.2 110195.6 114782.8 114905.9 109706.2
#> [137] 114909.4 114602.9 114577.5 114972.6 109954.4 114913.1 109480.6 113847.9
#> [145] 111743.0 114885.7 114885.2 112678.1 114818.3 114954.2 114648.8 109609.4
#> [153] 114639.9 108882.0 109614.8 114765.4 109125.0 114815.3 110702.0 109043.5
#> [161] 114969.8 114779.6 109444.8 113275.1 113232.1 113330.6 114443.3 114862.1
#> [169] 114862.9 110197.5 115082.1 114639.6 114760.8 114749.2 114727.8 113571.8
#> [177] 114790.6 115022.4 115019.5 114767.3 115103.2 114580.4 114804.5 114921.2
#> [185] 114892.4 114904.2 114764.7 114986.4 114873.9 114957.7 114840.0 109957.2
#> [193] 115049.8 114827.8 113570.8 114938.0 114669.2 114545.9 114900.1 114858.1
#> [201] 114911.2 115096.5 114842.9 114244.0 109986.2 114856.4 114835.2 114898.0
#> [209] 113884.6 114955.8 113820.2 114901.5 114194.3 114716.9 114834.0 114968.8
#> [217] 114354.5 115012.5 114847.4 115036.6 114961.8 111812.1 114664.3 114887.4
#> [225] 114782.4 109217.5 114851.3 114843.4 114611.9 111788.0 114982.2 114703.1
#> [233] 114871.1 114814.0 114817.0 114794.5 114165.6 114780.3 109296.5 114684.5
#> [241] 114994.1 114522.9 109158.0 115053.4 114789.9 115049.2 115084.5 114838.6
#> [249] 114829.0 110219.0 109117.8 109386.6 115062.0 111314.4 113904.2 114891.0
#> [257] 114832.3 110154.0 114869.7 114774.7 114875.7 114896.1 110061.8 115027.0
#> [265] 115068.9 114362.8 109431.3 109914.3 114890.7 114744.5 114932.6 114747.8
#> [273] 114732.1 114912.8 114811.8 115001.5 114843.2 114785.7 114774.1 115029.3
#> [281] 113719.7 114851.0 114345.3 115022.4 114983.5 114905.4 114976.2 114694.8
#> [289] 115017.0 114926.3 114917.1 109196.7 114871.3 114703.2 114898.3 114812.1
#> [297] 114933.5 114994.3 113824.7 110040.4 110108.2 114848.5 115013.3 114798.3
#> [305] 110548.5 114867.8 115052.0 114742.0 114789.2 113526.6 110163.5 110156.6
#> [313] 115032.8 114887.2 114789.5 114978.0 114942.6 110119.9 114641.0 109362.5
#> [321] 109475.5 114808.8 114898.6 113997.5 109145.7 114382.8 114848.5 108786.4
#> [329] 114747.7 109064.9 114953.9 114830.1 114245.5 109157.7 114799.0 114286.6
#> [337] 114471.5 114372.7 114242.4 114246.9 114218.8 114240.0 114280.1 114371.3
#> [345] 114392.6 114194.3 114199.2 114215.4 114707.1 114171.7 114038.0 109116.8
#> [353] 114957.3 113791.6 114268.9 114988.7 111615.7 115080.4 114318.2 109240.3
#> [361] 115122.8 114914.6 114890.4 113515.4 110032.1 114421.0 114853.1 115021.2
#> [369] 114848.6 114704.9 114781.4 114831.1 114959.6 114919.4 114785.3 114870.2
#> [377] 115069.3 114792.3 114890.1 114890.4 109605.8 109440.5 114496.2 109570.5
#> [385] 114895.0 114045.2 115008.0 114368.4 108879.7 114211.4 114761.0 114217.1
#> [393] 114834.3 115037.0 114813.1 114977.5 114883.1 114952.1 110268.2 113842.9
#> [401] 109074.4 113889.7 114591.4 114899.1 115015.6 114822.6 109589.5 114780.1
#> [409] 113553.9 114855.6 114639.3 110353.1 114725.3 114664.1 109522.5 109430.9
#> [417] 113567.0 113464.5 114702.9 114772.8 114385.3 114817.1 114775.9 113550.7
#> [425] 113508.4 114810.9 114898.8 110292.6 115077.6 115050.0 114807.1 115105.7
#> [433] 114606.2 114703.2 114206.8 110169.3 114886.6 114906.0 114854.1 109209.7
#> [441] 114796.0 109756.5 114996.6 114667.4 114676.0 114849.9 114867.0 114829.2
#> [449] 114639.5 114378.5 114863.6 109609.2 114759.7 114725.2 109509.8 115089.3
#> [457] 114814.8 114834.8 114930.0 114849.4 114929.2 114950.5 114979.8 113843.4
#> [465] 114910.4 109521.6 114863.6 114919.3 113752.7 114934.1 114956.3 114887.5
#> [473] 108956.6 114807.1 113536.9 112791.5 113974.1 114744.8 114801.8 114823.9
#> [481] 114784.7 114715.7 113602.4 114958.7 114656.7 113274.0 115010.5 114813.9
#> [489] 110311.6 113861.8 110025.2 114792.1 114871.5 114919.1 112612.4 114866.9
#> [497] 113833.0 114749.8 114899.7 114894.6 114694.1 114829.6 109336.3 114850.3
#> [505] 114950.1 109007.7 108956.7 108933.9 114714.9 114347.4 114868.3 114831.0
#> [513] 114818.9 115091.7 114875.5 108932.5 114808.3 114646.2 114628.2 114748.4
#> [521] 115044.9 114504.9 109654.1 113568.0 114784.0 109535.1 109384.3 114841.6
#> [529] 113419.1 109475.9 109391.3 114867.2 109478.2 114510.1 109539.3 114895.4
#> [537] 109354.9 109550.8 115114.8 109510.3 109466.3 109463.4 109408.7 109354.4
#> [545] 109497.8 109065.7 114895.3 114688.4 114868.9 114965.9 110373.5 109460.4
#> [553] 114574.6 114663.4 114656.2 114635.7 114582.4 114780.5 109219.4 115065.0
#> [561] 114670.1 113758.6 114774.8 114701.1 114892.5 114208.2 114954.7 114929.1
#> [569] 113250.7 109481.8 114772.2 114981.4 114729.2 115142.3 114829.1 115039.1
#> [577] 114854.1 115079.0 114322.4 109526.7 114928.7 109074.1 114933.5 114839.0
#> [585] 114846.7 110166.6 114869.3 114902.1 114953.6 114785.1 114842.5 109393.9
#> [593] 109133.4 114687.7 110694.0 109339.7 114945.7 114707.7 114669.4 114192.0
#> [601] 114213.8 114824.6 114862.7 112907.0 113458.4 114672.8 114744.8 114804.4
#> [609] 113437.9 113900.1 111584.4 109792.2 114987.8 114761.0 115089.0 114664.6
#> [617] 110229.9 115076.6 108994.9 113486.9 113339.5 114929.8 114809.1 111786.6
#> [625] 114808.5 113418.5 114730.7 112540.5 115095.5 114861.8 114044.2 114146.4
#> [633] 114894.8 114960.3 115100.4 115007.6 109487.9 114890.5 114771.0 114875.6
#> [641] 109674.6 114172.7 114750.7 114833.4 114562.8 112636.0 114855.2 113024.7
#> [649] 114895.8 115081.6 115105.4 108920.7 114842.8 110187.4 111955.1 114998.5
#> [657] 114240.1 114889.6 114909.4 113829.9 115059.0 115044.1 114673.9 114887.2
#> [665] 114865.8 114937.2 114840.0 114776.9 114951.2 114963.6 108989.3 115110.4
#> [673] 114936.5 109550.1 114553.5 114890.4 114763.9 114828.6 114950.3 110124.1
#> [681] 114838.0 114622.0 114988.1 114864.8 110039.5 114998.6 115013.9 114821.4
#> [689] 115078.2 114964.9 114985.2 114883.3 114816.4 114728.3 114824.5 115006.1
#> [697] 114912.3 109468.2 115106.4 109397.7 114456.4 114935.3 114804.2 114823.3
#> [705] 114879.9 114882.8 114799.4 114755.0 114874.4 114813.2 109621.8 114694.2
#> [713] 111095.1 114906.6 111450.3 114766.2 114464.1 114899.0 114996.1 111741.3
#> [721] 115013.3 114478.5 114575.3 109725.4 110146.3 115154.8 114908.7 114858.9
#> [729] 114768.0 115099.2 114498.5 113821.1 114742.8 109421.6 114950.3 114763.8
#> [737] 114821.8 114988.1 114657.5 114996.1 114815.9 114697.4 112595.0 114945.9
#> [745] 113788.5 113868.3 114698.7 114821.2 114576.1 115076.9 114690.1 114822.7
#> [753] 110658.8 114764.8 109383.2 114913.9 114705.2 109446.5 114865.6 114826.3
#> [761] 114869.7 114927.1 114845.1 109257.4 109495.4 114956.6 114861.2 109108.4
#> [769] 114770.2 114839.0 115051.9 114870.5 114791.5 114847.0 114801.9 114627.9
#> [777] 114917.6 108796.7 111584.4 111752.7 114580.0 114712.5 114840.2 114162.2
#> [785] 109319.4 114869.2 114651.4 114792.9 115054.0 114699.0 114990.2 114127.6
#> [793] 114781.1 113575.8 110180.7 114933.6 114642.3 115011.0 114654.9 114809.6
#> [801] 114615.8 110048.3 114839.7 114784.1 114837.3 114921.3 114908.2 114804.5
#> [809] 114811.4 114681.9 109549.6 114684.3 115166.8 114795.8 114884.4 114888.1
#> [817] 114764.5 114835.5 114739.9 114815.7 114864.1 114811.2 109974.2 114759.1
#> [825] 109423.7 114832.0 114860.7 114824.7 113747.2 114837.9 114939.1 114975.5
#> [833] 115008.2 114827.9 109772.1 114929.8 114740.9 110250.7 115118.9 114850.0
#> [841] 114431.7 110921.9 114557.9 115005.9 114631.5 114682.1 114864.5 114819.4
#> [849] 114805.0 109231.9 109560.0 114819.1 115041.7 114923.6 110353.2 114857.3
#> [857] 113892.1 109335.5 114896.2 114829.2 115133.1 114906.3 109127.3 110261.7
#> [865] 114861.5 114874.4 113035.2 109268.7 114911.5 109319.9 109298.9 114774.8
#> [873] 115058.0 115031.7 114999.3 114823.0 114796.3 115054.1 114939.0 114765.8
#> [881] 109100.6 109175.2 114358.6 114904.9 110258.8 110936.2 114629.6 112788.2
#> [889] 109438.0 114788.6 109349.1 114806.1 115023.7 114711.6 114627.8 114801.2
#> [897] 114774.5 114923.7 113542.1 114847.2 113503.2 114941.6 115031.1 115027.8
#> [905] 114841.1 111167.2 110260.5 114064.2 114857.7 111765.6 114838.0 114884.3
#> [913] 109027.4 110928.2 114828.6 114650.8 114603.9 115059.8 109671.7 114880.9
#> [921] 114996.9 114977.4 114300.6 114756.5 109407.9 114029.0 114409.2 109189.2
#> [929] 114978.7 114826.7 114757.9 114826.6 114906.7 114855.5 115024.1 114736.6
#> [937] 114894.3 115053.4 114824.6 114209.8 115004.9 114586.2 114610.7 113809.7
#> [945] 115017.1 114209.8 114608.5 114602.6 115010.0 113108.6 114920.8 114844.0
#> [953] 114821.9 114825.3 114294.8 114771.8 114856.0 114788.5 114501.2 109557.1
#> [961] 109277.8 114887.9 114720.5 115019.1 114787.2 114751.4 109417.9 114486.7
#> [969] 114527.2 114453.4 114688.1 109953.3 114671.0 114949.4 114819.6 114112.9
#> [977] 115074.7 111512.0 114847.9 111525.4 114210.4 114151.2 114754.7 109356.7
#> [985] 111809.3 111328.1 114996.9 114895.4 114913.6 114666.1 114878.8 114930.6
#> [993] 114752.2 114552.0 109344.8 114791.4 114804.2 111124.0
#> 
#> 
#> [[3]]
#> [[3]]$residus
#> [1] 1712237957   88327926   31177321   37189590   29075426   28572303   27496091
#> [8]   39932545   48119142
#> 
#> [[3]]$delta.star.id
#> [1] 7
#> 
#> [[3]]$hat.X.delta.star
#>   [1] 16831.06 16795.30 16272.44 17029.67 16731.96 16854.23 16374.68 16180.33
#>   [9] 15938.66 17109.84 16576.72 16410.88 16967.67 16654.57 16393.64 17006.69
#>  [17] 16470.39 17079.40 16367.78 16373.38 16789.25 16456.78 16455.12 15975.31
#>  [25] 16891.75 17118.50 16442.84 17101.12 16777.17 16837.81 16412.20 16296.58
#>  [33] 16983.59 16295.61 16896.92 16131.86 16354.55 17065.76 17035.73 16295.42
#>  [41] 15548.91 16880.95 16217.85 16965.80 16345.24 15960.83 16192.00 16409.80
#>  [49] 16250.79 16838.49 16939.97 16831.83 16442.83 16632.35 16784.21 16870.87
#>  [57] 16518.43 16573.24 17034.92 16234.05 16728.45 16725.69 16520.27 16380.91
#>  [65] 16810.32 17045.19 16840.75 16515.01 16887.03 16963.00 16983.30 16842.92
#>  [73] 16451.45 16342.12 16961.92 16330.82 16774.03 16403.83 16631.55 16077.19
#>  [81] 15844.78 16371.08 16859.93 16840.86 17167.48 16735.50 16343.33 16323.31
#>  [89] 16891.23 16991.81 16384.13 16358.86 16709.68 16368.68 16867.22 16808.82
#>  [97] 16160.87 16976.28 16436.24 16183.99 16961.47 16058.79 16524.62 15869.43
#> [105] 16609.13 16860.97 15685.97 16353.02 16567.91 15505.69 16993.83 17024.48
#> [113] 16887.09 16098.86 16105.43 16178.30 16536.86 16469.96 16326.47 16325.21
#> [121] 16319.20 16723.50 16528.44 16819.49 15935.57 16263.27 16880.42 16027.19
#> [129] 16117.43 16869.39 16947.08 17029.92 16323.48 16304.72 16649.00 16286.95
#> [137] 16859.08 16307.00 16287.29 16956.26 16441.51 16738.83 16393.96 16003.75
#> [145] 16675.81 16692.65 16935.39 16443.67 16342.44 17068.94 16588.82 16328.19
#> [153] 16243.44 16413.56 16331.42 16738.01 16391.48 16258.31 16353.68 16411.57
#> [161] 17020.31 16190.82 16416.41 15694.75 15560.02 15653.68 16112.74 16841.90
#> [169] 16397.04 16397.70 16999.01 16178.37 16831.68 16980.22 17026.62 15841.79
#> [177] 16485.68 17009.13 17033.29 16901.74 17057.27 16206.99 16711.99 16728.12
#> [185] 16808.07 16882.53 16581.87 16815.08 16320.64 16964.69 16807.22 16408.17
#> [193] 17041.20 16292.79 15729.15 16807.35 16405.49 16389.80 16787.99 17051.48
#> [201] 16826.26 17042.08 16967.52 16306.33 16385.51 16638.83 16557.51 16844.84
#> [209] 15952.05 16818.40 15875.99 16969.29 16078.33 16700.28 16420.05 16684.48
#> [217] 16263.12 17066.20 16786.69 17060.66 16888.93 16364.23 16536.02 16791.02
#> [225] 16255.19 16368.55 16666.34 16855.77 16380.48 16389.75 16939.88 16159.52
#> [233] 16998.53 16817.73 16239.82 16348.88 16155.90 16536.92 16316.66 16235.74
#> [241] 16799.39 16382.07 16316.86 17027.22 16970.17 17077.60 17082.20 16846.28
#> [249] 16417.90 16396.42 16460.97 16279.42 17091.03 16298.33 16067.06 16838.80
#> [257] 16988.37 16430.64 16332.16 16350.31 16822.94 16758.46 16416.95 17056.42
#> [265] 16862.44 16245.38 16355.84 16241.87 16829.47 16729.28 16884.79 16598.87
#> [273] 16301.70 16828.27 16542.52 17033.52 16379.16 16264.32 16567.52 16908.20
#> [281] 15814.30 16553.98 17077.31 16944.57 16965.75 16297.98 16888.74 16556.17
#> [289] 16988.89 16879.08 16966.46 16379.66 16781.10 16206.01 17069.30 16914.52
#> [297] 16152.26 17062.43 15906.79 16467.92 16383.11 16742.79 16979.12 16652.53
#> [305] 16491.01 16387.12 17063.94 16517.07 16695.22 15916.98 16460.05 16357.39
#> [313] 17048.68 16664.63 16669.64 17039.37 17176.70 16405.25 16079.52 16441.42
#> [321] 16416.15 16637.96 16821.20 16134.96 16308.19 16267.04 16712.66 16445.91
#> [329] 16601.46 16370.74 16912.74 16297.36 16126.81 16418.14 16437.52 16091.60
#> [337] 16216.71 16118.79 16152.63 16113.15 16229.90 16138.38 16145.94 16184.26
#> [345] 16146.80 16134.22 16147.93 16178.99 16577.24 16177.85 16031.20 16379.09
#> [353] 17031.70 15986.74 16177.80 17144.35 16346.73 17034.81 16158.12 16411.84
#> [361] 17045.56 16962.23 16534.82 15787.97 16294.29 16116.75 16808.61 17043.49
#> [369] 17079.37 16420.17 16354.25 16798.81 17059.65 16818.26 16710.05 16862.94
#> [377] 16990.38 16955.20 16970.79 16990.90 16339.52 16374.50 16248.08 16349.43
#> [385] 16861.94 16018.54 17052.59 16175.48 16479.15 16103.91 16718.31 16157.42
#> [393] 16645.98 17157.56 16243.34 17078.26 16725.16 17056.97 16420.72 15975.50
#> [401] 16414.83 15986.61 16052.30 16985.79 16988.38 16761.52 16359.87 16324.26
#> [409] 15818.27 16872.68 16587.13 16397.38 16149.61 16628.96 16311.72 16387.26
#> [417] 15856.15 15790.60 16696.86 16329.29 16190.12 16910.20 16585.90 15729.98
#> [425] 15860.94 16740.19 16735.00 16356.83 17047.46 16999.31 16362.81 16952.59
#> [433] 16095.02 16619.30 16090.81 16320.81 16778.53 16803.65 17033.83 16386.07
#> [441] 16700.33 16336.12 16790.57 16513.05 16634.81 16941.23 16828.88 16645.65
#> [449] 16065.60 16170.79 16279.13 16311.78 16499.27 16764.67 16409.56 16988.67
#> [457] 16982.68 16402.23 17027.37 16988.58 17138.40 16988.51 17041.01 15958.81
#> [465] 16650.25 16374.29 17016.66 17043.10 15820.87 16867.54 17023.29 16934.18
#> [473] 16413.32 16705.70 15724.56 15426.75 16056.43 16360.39 16801.31 16631.90
#> [481] 16997.24 16490.70 15713.16 16984.42 16439.59 15760.79 17133.69 16754.62
#> [489] 16482.15 15889.13 16399.12 16303.16 16930.32 16926.91 16334.34 16907.46
#> [497] 15880.09 16295.01 16831.03 16834.45 16724.58 16631.15 16419.38 16315.02
#> [505] 16965.53 16368.84 16359.55 16427.14 16116.21 16134.83 16698.47 16832.56
#> [513] 16419.75 17135.19 16901.36 16464.96 16834.37 16768.30 16155.69 16699.34
#> [521] 16921.30 16205.13 16371.94 15808.55 16665.07 16375.57 16422.95 17121.34
#> [529] 15729.47 16371.46 16324.68 16774.22 16404.80 16502.24 16302.96 17063.38
#> [537] 16387.08 16374.66 17081.18 16398.28 16384.06 16446.07 16413.73 16430.51
#> [545] 16258.40 16476.83 16874.53 16847.42 16645.22 16811.19 16258.84 16293.16
#> [553] 16359.68 16401.51 16417.85 16466.84 16468.17 16539.48 16402.12 17014.49
#> [561] 16295.37 15887.24 16625.78 16133.92 16777.87 16042.86 17005.62 16758.36
#> [569] 15698.47 16379.92 16850.10 17023.37 16678.25 16984.84 16889.44 17060.28
#> [577] 16763.09 17047.99 16198.21 16344.97 16852.02 16405.03 16838.28 16750.40
#> [585] 16980.11 16280.50 16172.21 16657.05 16365.80 16788.47 16668.65 16415.47
#> [593] 16437.86 16559.33 16375.80 16506.21 17047.02 16343.23 16358.65 16097.76
#> [601] 16174.40 16593.51 16677.21 16592.88 15778.27 16615.03 16632.18 16266.44
#> [609] 15730.93 15958.17 16437.87 16332.19 16866.09 16954.95 17077.04 16523.92
#> [617] 16391.90 16991.40 16474.86 15779.95 15688.67 16257.12 16354.24 16487.45
#> [625] 16337.28 15846.07 16362.58 16418.72 16846.04 16562.30 16072.44 16084.04
#> [633] 16792.57 16743.24 16980.26 16888.90 16413.08 16843.81 16312.09 16647.00
#> [641] 16325.75 16121.63 16714.98 16748.26 16389.85 15537.19 16979.31 15519.72
#> [649] 16725.19 17100.09 17050.15 16288.48 16401.38 16410.53 16358.19 17117.32
#> [657] 16120.97 17055.43 17074.23 16030.49 16889.47 17018.38 16408.46 16974.76
#> [665] 16995.57 17008.44 16932.76 16336.78 16869.10 17139.72 16459.82 16990.11
#> [673] 16907.56 16390.78 16297.67 16954.16 16343.22 16406.62 16811.68 16285.57
#> [681] 16782.50 16248.86 16891.00 16959.49 16361.20 17006.04 17098.81 16738.07
#> [689] 17093.97 17096.76 16935.48 16774.91 16726.25 16139.48 16324.04 17034.88
#> [697] 17028.73 16293.59 16996.62 16408.35 17136.72 17062.90 16182.02 16761.77
#> [705] 16874.99 16397.50 16848.03 16636.26 16826.11 16961.74 16366.44 16639.10
#> [713] 16445.46 16522.48 16606.88 16712.57 16242.35 16831.98 17030.31 16377.65
#> [721] 17091.07 16002.40 16127.18 16316.20 16414.35 17029.49 16872.44 16380.06
#> [729] 16642.21 17044.25 16166.34 15825.54 16132.50 16389.23 16863.69 16544.10
#> [737] 16154.98 16742.26 16269.38 16651.22 16844.75 16603.70 15414.21 16958.74
#> [745] 15799.08 15865.95 16513.18 16603.88 16136.35 16981.00 16579.77 16751.05
#> [753] 16396.17 16175.03 16346.21 16917.09 16617.16 16417.45 16872.81 16712.46
#> [761] 16663.59 16943.34 16648.90 16370.75 16398.21 16739.30 16712.10 16406.63
#> [769] 16597.12 16675.93 17072.91 16727.51 16604.22 16531.71 16510.96 16635.87
#> [777] 16714.47 16446.70 16369.16 16334.91 16335.20 16722.33 16899.75 16197.62
#> [785] 16416.90 16349.85 16632.56 16780.71 17077.82 16157.93 16737.16 16165.60
#> [793] 16650.06 15800.63 16354.10 16824.22 16465.93 17083.89 16401.60 16318.73
#> [801] 16431.61 16406.42 16729.01 16802.80 16662.60 16827.23 16626.36 16210.44
#> [809] 16562.94 16845.75 16270.67 16417.96 17054.06 16575.16 17060.34 16928.28
#> [817] 16558.12 16476.85 16189.60 16723.06 16507.12 16734.29 16311.74 16630.76
#> [825] 16306.94 16949.52 16901.88 16667.38 15865.77 17026.71 16882.13 16844.08
#> [833] 16983.81 16954.52 16298.36 16262.58 16669.01 16326.91 17023.93 16917.26
#> [841] 16140.10 16302.00 16078.99 17015.74 16183.79 16110.55 16448.88 16644.80
#> [849] 16954.33 16476.33 16336.77 16819.38 17106.78 16831.53 16398.73 16984.36
#> [857] 15964.62 16517.76 16668.72 16724.77 16985.96 16656.12 16429.00 16353.90
#> [865] 16484.67 16969.81 16511.32 16433.19 16698.96 16427.89 16379.44 16515.57
#> [873] 17080.07 16919.35 16982.60 16540.50 16488.54 16998.28 16933.16 16676.43
#> [881] 16379.37 16457.68 16126.36 16260.72 16347.67 16321.22 16443.43 15491.27
#> [889] 16364.72 16553.92 16455.32 16603.30 17116.19 16285.25 16642.25 16482.85
#> [897] 16964.35 16983.96 15787.52 16770.68 15811.17 16805.44 17095.86 17139.64
#> [905] 16245.00 16394.83 16369.72 16074.81 16405.64 16407.98 16945.21 16693.77
#> [913] 16457.20 16289.01 16776.44 16448.07 16217.36 17009.12 16385.94 16690.80
#> [921] 17046.38 17015.97 16116.71 16681.22 16444.28 16050.17 16157.60 16386.11
#> [929] 16872.04 16338.65 16305.75 16731.75 16730.51 16776.44 16908.78 16543.84
#> [937] 16677.14 17003.72 16735.31 16060.92 16902.06 16398.41 16377.39 15993.94
#> [945] 16971.65 16157.98 16530.54 16394.72 16992.48 15595.92 16858.82 16294.63
#> [953] 16520.89 16296.95 16240.42 16645.32 16655.80 16651.07 16369.24 16401.96
#> [961] 16462.00 16875.84 16579.02 17012.99 16363.28 16661.15 16304.25 16369.61
#> [969] 16248.35 16322.79 16439.84 16329.06 16478.38 16872.30 16717.95 16081.79
#> [977] 17017.59 16307.90 16765.33 16378.23 16087.77 16174.15 16949.34 16402.78
#> [985] 16317.20 16438.56 17014.38 17072.12 16799.96 17009.40 16820.49 17095.61
#> [993] 16841.03 16107.72 16404.08 16813.80 16894.39 16339.29
#> 
#>