matchingVariograms

Function to create a permutation of spatial dataset X by starting with a randomly shuffled version of X and then smoothing it until the variogram of this permutation matches the variogram of the unmodified X

Usage

matchingVariograms(
  X.randomized,
  long,
  lat,
  delta,
  target_variog,
  prctile,
  ids,
  i
)

Arguments

X.randomized

numeric vector random permutation of the values of X across its x,y-coordinates.

long

numeric vector: y-coordinates of points in dataset X

lat

numeric vector: x-coordinates of points in dataset X

delta

numeric vector Given a point in dataset X, the percentage of neighbors which should be within the smoothing kernel. This can be a list or a single numeric. Input to `locfit::lp` as `nn`

target_variog

variogram variogram of dataset X that will be used as target when doing the matching. See geoR::variog

prctile

numeric: the max distance that should be used for evaluating the fit between the variograms from the permuted data and the target variogram. Can be determined from calculating all distances between the points and then calculating a specified quantile. Input to `geoR::variog` as `max.dist`

ids

integer: indices from 1 to the number of points in the dataset X

i

integer the number indicating the ith permutation

Value

The output is returned as a list.

residusnumeric vector of length of delta, sum of squares of the residuals between the target variogram and the permuation variogram for each delta

delta.star.idinteger, the index of the delta that minimizes the residual sum of squares

hat.X.delta.starnumeric vector of length of X.randomized, permutation generated from optimal delta, whose variogram best matches the variogram of X

References

adapted from: Viladomat, Júlia et al. “Assessing the significance of global and local correlations under spatial autocorrelation: a nonparametric approach.” Biometrics vol. 70,2 (2014): 409-18. doi:10.1111/biom.12139

Examples

data(quakes)

#remove duplicated positions
quakes_data <- quakes[!duplicated(cbind(quakes$lat, quakes$long)),]

lat <- quakes_data$lat
long <- quakes_data$long
X <- quakes_data$depth
ids <- 1:length(X)

#number of permutations
B <- 3

#sequence of deltas
delta <- seq(0.1,0.9,0.1)

# maximum distance for the variogram set at the 25% percentile of
# the distribution of pairs of distances:
dists <- dist(cbind(lat, long))
prctile <- quantile(dists, probs = 0.25)

# variogram of variable X that will be used as target when doing the matching:
target_variog <- geoR::variog(data = X, coords = cbind(long, lat),
                              max.dist = prctile, option = "bin",
                              messages = FALSE)

# create B random permutations of the values of X across locations:
X.randomized <- lapply(1:B, function(i) {
   sample(X, size = length(X), replace = FALSE)
})

# create B permutations of the values of X with variograms that match the
#target variogram of X:
output <- lapply(1:B, function(i) {
  output <- matchingVariograms(X.randomized[[i]], long, lat,
                               delta, target_variog, prctile, ids, i)
})

output
#> [[1]]
#> [[1]]$residus
#> [1] 886792441  80487344   8348616  26934936  15085992  15258405  28850761
#> [8]  21129226   7159215
#> 
#> [[1]]$delta.star.id
#> [1] 9
#> 
#> [[1]]$hat.X.delta.star
#>   [1] 26531.71 26599.82 26686.57 26357.95 26448.99 26207.49 27669.72 26751.73
#>   [9] 26788.47 26454.37 26677.53 27634.82 26341.25 26607.30 27597.74 26017.25
#>  [17] 27717.09 26252.12 27000.95 26824.62 26654.88 27649.13 27014.56 26727.82
#>  [25] 26577.89 26315.84 27604.94 26420.01 26517.06 26338.22 26965.02 27743.01
#>  [33] 26002.57 26890.19 25989.52 27103.54 27711.32 26324.12 26366.86 27679.18
#>  [41] 26722.97 25892.85 27044.09 25762.99 27612.33 26807.34 26903.64 27659.89
#>  [49] 26948.50 26128.56 26261.43 26277.74 27479.79 26733.16 26383.47 26463.82
#>  [57] 26802.79 26796.27 26304.72 27124.75 26479.35 26497.13 27258.37 27613.29
#>  [65] 26607.53 26014.04 26515.12 26868.58 26454.47 25891.76 25887.42 26347.31
#>  [73] 27672.73 26996.13 26239.87 26968.11 26200.99 27560.47 26651.47 26790.45
#>  [81] 26734.94 26868.25 26419.02 26436.13 26220.09 26169.47 27627.23 26970.11
#>  [89] 26006.38 25932.70 27646.91 27632.75 26454.63 27678.76 25961.46 26488.55
#>  [97] 26873.37 26219.76 27595.55 26971.15 26012.16 27114.00 26822.32 26761.91
#> [105] 26639.21 26321.12 26627.73 27562.75 26422.51 26655.07 25899.79 26154.21
#> [113] 26308.14 27136.45 27126.97 27051.60 27122.84 27187.85 27577.54 27633.88
#> [121] 27678.76 26649.28 26782.80 26326.98 26825.43 27709.80 26294.72 27035.53
#> [129] 26855.47 26235.52 25975.78 26279.96 27657.87 27002.15 26724.87 27615.75
#> [137] 25871.57 26814.66 26840.57 26232.45 27417.15 26601.69 27618.61 26723.47
#> [145] 27073.50 26249.94 26112.91 27363.05 26964.97 26289.93 26611.84 27606.81
#> [153] 26718.70 27736.19 27590.00 26561.07 27597.47 27081.24 27626.23 27612.53
#> [161] 26532.67 27114.74 27608.73 26706.05 26645.27 26831.62 26797.09 26273.02
#> [169] 26866.40 27650.12 26301.77 26990.07 26575.04 26020.93 26123.56 26756.88
#> [177] 26749.36 26056.87 26253.46 26014.32 26263.49 26675.19 26336.21 26528.35
#> [185] 26541.31 26114.16 26658.02 26374.52 27083.60 25959.12 26360.78 27666.23
#> [193] 26278.79 27090.24 26766.53 26628.25 26731.58 26692.93 26429.21 26365.57
#> [201] 26415.49 26242.76 26069.44 26756.01 27412.09 26621.48 26745.88 26500.04
#> [209] 26827.53 26435.98 26846.05 25939.88 26794.87 26352.88 26939.18 26512.22
#> [217] 26816.76 26312.78 26359.87 26143.22 26057.66 27608.53 26466.06 26391.75
#> [225] 26901.89 27586.42 26423.70 26009.20 26774.96 27552.50 26520.70 27003.97
#> [233] 26077.74 26234.99 26956.10 26903.72 26716.92 26771.57 27723.49 26968.49
#> [241] 26406.89 26694.40 27715.00 26187.69 26046.13 26314.54 26289.27 25889.83
#> [249] 26932.27 27648.21 27565.85 27576.29 26284.05 27716.28 26860.93 26605.47
#> [257] 26101.13 27661.18 27103.26 26972.46 26217.52 26506.94 27653.39 26338.48
#> [265] 26316.28 26678.40 27703.81 27727.39 26494.52 26648.98 26530.90 26810.39
#> [273] 26928.41 26393.55 26766.91 26374.95 26953.52 27157.08 26779.68 26290.49
#> [281] 26712.76 26757.19 26648.46 26069.63 26106.98 26984.80 26548.75 26283.85
#> [289] 26342.73 26497.71 26368.27 27649.79 26673.34 27033.39 25833.37 26010.42
#> [297] 27226.92 26370.45 26719.57 27544.42 27493.06 26321.68 25860.35 26610.71
#> [305] 27398.42 26900.62 26294.16 26799.59 26707.52 26725.43 27307.59 27377.05
#> [313] 26372.01 26608.55 26501.81 26326.28 26280.98 27715.38 26990.97 27533.61
#> [321] 27662.31 26441.97 26450.78 26681.05 27713.52 26899.12 26648.32 27763.13
#> [329] 26226.25 27693.14 26291.71 26917.10 26786.45 27702.50 27021.86 26781.99
#> [337] 26815.61 26781.42 26869.02 26862.37 26745.22 26879.39 26758.50 26820.97
#> [345] 26748.79 26778.31 26755.04 26813.10 26416.81 26889.22 26689.02 27613.08
#> [353] 26276.75 26724.14 26804.94 26384.59 27600.84 26133.76 26824.99 27576.07
#> [361] 26324.43 26000.46 26852.71 26697.78 27656.50 26735.67 26473.16 26189.02
#> [369] 26155.49 26622.15 26925.74 26330.49 26320.98 26469.78 26348.42 25951.15
#> [377] 26237.64 25901.77 25907.54 25941.85 27585.92 27574.53 26695.80 27604.02
#> [385] 26536.75 26874.06 26259.92 26946.53 27832.78 26629.95 26225.61 26827.72
#> [393] 26635.34 26178.17 26969.25 26270.38 25869.46 26329.11 27376.71 26691.36
#> [401] 27684.03 26702.70 26956.86 26005.13 26228.06 26495.33 27601.72 26963.29
#> [409] 26725.85 25982.92 26513.50 27691.05 27138.25 26355.14 27550.11 27640.90
#> [417] 26699.26 26730.96 26163.97 26869.42 26819.68 25957.49 26819.19 26776.56
#> [425] 26741.20 26607.24 26440.70 27637.75 26397.01 26269.10 26858.29 26337.36
#> [433] 27001.67 26545.65 26720.56 27709.37 26493.40 26494.64 26052.59 27567.56
#> [441] 26342.73 27649.65 26331.47 26533.18 26517.65 25986.14 26564.64 26640.57
#> [449] 27047.90 26732.08 27089.00 27678.24 26713.32 26315.54 27520.67 26308.06
#> [457] 26101.57 26987.78 26476.62 25915.76 26450.00 26402.38 26396.23 26885.74
#> [465] 26515.41 27636.92 25994.83 26423.38 26785.10 26212.91 26089.80 26032.82
#> [473] 27708.11 26276.08 26774.27 26651.38 26652.25 26915.24 26235.07 26638.73
#> [481] 25859.87 26539.98 26772.03 26290.03 26660.50 26696.73 26260.56 26585.91
#> [489] 27308.27 26763.80 27675.15 26964.81 26019.18 25924.05 27535.32 25910.75
#> [497] 26821.45 27002.42 26221.63 25909.43 26231.68 26417.92 27658.53 26957.74
#> [505] 26346.15 27688.62 27645.58 27744.75 26979.45 26740.89 26388.82 25874.19
#> [513] 27025.96 26184.72 25822.21 27678.35 26402.14 26197.01 26984.84 26296.98
#> [521] 26281.03 26842.72 27658.64 26720.56 26702.32 27683.68 27654.21 26047.18
#> [529] 26734.96 27618.43 27629.36 26495.45 27634.15 26532.71 27655.84 26265.86
#> [537] 27734.58 27722.93 26221.92 27746.49 27594.18 27751.88 27597.73 27702.98
#> [545] 27600.40 27699.29 26453.05 25840.52 26571.73 26542.17 27618.76 27706.18
#> [553] 26557.36 26514.29 26562.59 26619.51 26532.80 26713.94 27705.58 26378.08
#> [561] 26605.18 26694.88 26449.38 27022.30 26280.26 26724.60 26335.10 26252.49
#> [569] 26700.91 27576.21 26039.32 26285.11 26198.29 26212.46 25954.84 26299.55
#> [577] 26334.49 26271.80 26650.35 27640.42 26639.04 27706.91 26453.15 26565.79
#> [585] 25913.30 27681.16 27007.53 26543.40 26905.55 26653.75 26584.02 27591.60
#> [593] 27602.64 26403.80 27692.87 27626.83 26037.13 26693.41 26484.26 26804.16
#> [601] 26762.64 26623.59 26753.23 27219.31 26658.51 26366.78 26458.83 26932.14
#> [609] 26726.97 26702.22 27523.35 27609.28 26399.85 25871.75 26263.29 26409.34
#> [617] 27693.00 26411.18 27710.95 26671.55 26752.65 27077.46 26981.43 27603.92
#> [625] 26989.38 26693.61 27004.37 27499.32 26254.08 26716.93 26707.19 26677.07
#> [633] 26680.90 26293.87 26315.21 26234.31 27568.74 26005.97 26922.84 26631.78
#> [641] 27674.62 26679.03 26328.73 26551.96 26582.89 26753.63 25867.35 26759.67
#> [649] 26494.10 26354.83 26226.50 27529.90 26901.42 27508.50 27611.36 26361.36
#> [657] 26824.07 26305.11 26396.09 26826.29 26302.22 26400.31 26908.06 26433.93
#> [665] 26359.94 26359.46 26025.11 26937.56 26193.54 26279.50 27688.41 26263.97
#> [673] 26267.81 27606.23 26808.91 25948.18 26879.58 27007.97 26556.02 27653.16
#> [681] 26305.59 26860.51 26324.37 25891.64 27644.08 26464.96 26145.04 26562.45
#> [689] 26283.27 26243.83 26172.19 26417.68 26546.48 26908.53 26943.49 26357.58
#> [697] 26039.44 27576.49 26305.71 27565.01 26726.21 26357.05 27027.10 26270.95
#> [705] 25879.65 27036.32 25938.78 26503.43 26555.60 25870.07 27700.65 26394.27
#> [713] 27637.24 26714.93 27007.02 26681.63 26795.10 26117.33 26127.61 27572.12
#> [721] 26296.63 27067.88 27071.09 27664.95 27612.92 26194.46 26262.29 26951.38
#> [729] 26637.12 26328.95 26985.51 26829.84 27044.56 27592.65 25936.15 26475.14
#> [737] 27147.49 26457.85 26948.17 26637.64 26113.37 26384.32 26683.51 26291.43
#> [745] 26666.00 26697.32 26560.45 26740.37 27002.54 26235.53 26379.63 26627.60
#> [753] 27606.21 27063.19 27566.79 25938.90 26381.15 27551.72 25953.00 26409.71
#> [761] 26683.89 26106.48 26427.62 27666.01 27589.96 26193.35 26291.53 27725.61
#> [769] 26272.24 26254.41 26189.93 26261.63 26676.20 26812.43 26631.35 26540.62
#> [777] 26690.08 27732.60 27573.74 27480.95 26613.67 26235.39 25961.73 26686.17
#> [785] 27595.31 26927.42 26329.69 26596.73 26280.33 27140.28 26420.11 26768.75
#> [793] 26604.82 26686.76 27628.09 26485.18 26519.99 26321.87 26609.33 26928.39
#> [801] 26571.42 27366.41 26677.96 26187.22 26627.37 26423.10 26650.40 26950.25
#> [809] 26659.13 26184.46 27726.70 26655.19 26336.63 26808.28 26026.76 25930.42
#> [817] 26797.63 26796.06 27071.13 26082.05 26838.36 26510.77 27680.43 26488.07
#> [825] 27597.98 25952.99 26016.14 26483.11 26737.37 25984.28 26137.26 26308.23
#> [833] 26348.77 25905.32 27706.52 26941.38 26388.60 27646.79 26287.92 25960.76
#> [841] 26792.23 27603.00 27033.49 26217.67 26959.62 27077.93 26697.30 26604.28
#> [849] 25843.11 27698.75 27728.81 26198.15 26337.15 26122.41 27427.03 25920.11
#> [857] 26837.64 27541.85 26678.70 26628.45 26242.00 26546.54 27643.75 27737.04
#> [865] 26691.62 25958.79 27352.86 27548.72 26516.82 27549.45 27480.00 26744.64
#> [873] 26401.59 26359.78 26239.04 26699.07 26935.34 26234.48 26160.64 26547.96
#> [881] 27650.25 27712.48 26703.33 27224.48 27709.50 27682.30 26518.91 26688.60
#> [889] 27659.82 26763.27 27607.15 26416.58 26343.17 26945.88 26350.72 26768.89
#> [897] 26018.26 26394.83 26745.72 26161.72 26732.25 26488.49 26340.46 26307.29
#> [905] 27125.21 27602.98 27680.54 26699.49 26879.15 27618.78 25837.49 26476.18
#> [913] 27726.36 27635.97 26272.08 26673.08 26757.48 26236.52 27668.59 26282.96
#> [921] 26273.83 26338.05 26760.35 26551.53 27695.75 26758.12 26843.72 27725.36
#> [929] 26050.41 26872.25 27008.86 26302.60 26301.96 26181.81 26307.34 26485.72
#> [937] 26618.26 26342.73 26539.97 26763.74 26372.02 26526.02 26623.74 26734.44
#> [945] 26285.57 26717.60 26692.14 26625.48 26390.47 26756.83 26132.17 27012.53
#> [953] 26949.91 26978.91 26710.09 26350.36 26638.67 26546.62 26621.05 27624.18
#> [961] 27624.33 26337.68 26547.31 26382.53 26921.52 26766.79 27586.43 26620.27
#> [969] 26823.19 26636.84 26640.85 27660.66 26601.76 26014.73 26247.27 26742.99
#> [977] 26453.80 27589.51 26561.86 27564.26 26703.04 26656.08 25869.32 27611.23
#> [985] 27576.20 27593.22 26294.62 25994.63 26452.99 26298.27 26293.09 26335.66
#> [993] 25996.35 27020.70 27661.24 26292.49 25963.98 27548.32
#> 
#> 
#> [[2]]
#> [[2]]$residus
#> [1] 520448662 261899057 246995290 271610869 222739060 205570085  83106388
#> [8]  74325461  42963032
#> 
#> [[2]]$delta.star.id
#> [1] 9
#> 
#> [[2]]$hat.X.delta.star
#>   [1] 114944.8 114878.9 114391.6 115000.5 114987.9 114901.7 109015.4 114235.8
#>   [9] 114012.3 114832.9 114745.1 109430.7 115101.3 114800.8 110767.4 114842.8
#>  [17] 109095.8 115026.7 114874.5 114812.6 114927.5 108983.8 114746.1 113942.4
#>  [25] 114965.4 115023.5 109540.8 114775.8 114811.1 114985.8 114910.1 109393.3
#>  [33] 114845.5 114815.6 114857.1 114638.9 110073.2 114984.6 115016.9 110094.8
#>  [41] 112841.3 114822.8 114759.9 114834.0 110256.1 113903.5 114516.1 109075.8
#>  [49] 114912.9 114885.7 115059.4 114825.9 112858.4 114840.5 114949.4 114949.8
#>  [57] 114703.5 114815.9 115059.2 114774.2 114911.7 114907.9 111173.0 109057.1
#>  [65] 114888.9 114903.7 114961.6 114865.6 114916.7 114899.1 114879.0 114881.9
#>  [73] 109049.9 114763.3 115005.0 114815.9 114871.3 109422.0 114886.1 114068.1
#>  [81] 113399.7 114781.5 114931.5 114903.4 115075.7 114833.8 111161.0 114832.5
#>  [89] 114852.5 114875.0 109511.6 109376.2 114835.5 110365.4 114797.3 115019.2
#>  [97] 114240.2 115068.1 109365.6 114782.3 114930.1 114575.8 114841.4 113522.0
#> [105] 114858.9 114988.8 113328.1 110496.8 114637.3 112986.2 114865.4 114995.9
#> [113] 115040.8 114601.7 114607.1 114663.5 111034.4 111044.3 109460.2 110098.7
#> [121] 109435.9 114665.6 114917.6 115001.1 113601.1 109708.1 114915.9 114507.8
#> [129] 114216.1 115018.4 114877.6 114870.9 110182.8 114703.5 114855.1 109578.1
#> [137] 114802.4 114601.1 114509.8 114924.2 110161.2 114945.0 109364.6 114009.6
#> [145] 111801.9 114817.2 114942.9 112722.5 114902.5 115062.6 114644.6 109645.2
#> [153] 114684.8 108967.3 109539.0 114785.2 108971.2 114823.5 110762.9 109046.4
#> [161] 114928.1 114834.8 109458.6 113253.1 113324.2 113412.7 114470.3 114942.7
#> [169] 114764.6 110164.3 115034.0 114654.2 114810.7 114754.2 114708.3 113550.4
#> [177] 114817.9 114957.7 115132.6 114786.4 115023.7 114495.0 114730.8 114860.6
#> [185] 114797.0 114779.8 114827.9 114944.0 114821.6 114736.7 114817.0 110081.1
#> [193] 115109.4 114820.0 113536.4 114945.4 114611.6 114708.2 114970.3 114891.7
#> [201] 114905.9 115098.6 114896.2 114252.0 109992.0 114835.5 114796.5 114888.4
#> [209] 113845.1 114918.7 113723.5 114870.3 114121.7 114638.2 114808.1 114848.5
#> [217] 114329.0 114960.7 114868.4 115018.0 115003.7 111797.7 114650.4 114952.5
#> [225] 114834.8 109299.9 114791.0 114860.5 114696.1 111738.5 114892.4 114689.7
#> [233] 114873.4 114791.3 114714.6 114785.6 114265.3 114793.3 109353.1 114730.2
#> [241] 115005.1 114558.5 109175.2 115042.4 114894.2 115051.0 115019.0 114831.8
#> [249] 114819.0 110259.2 109008.4 109495.8 114967.2 111332.7 113962.9 114939.5
#> [257] 114931.5 110249.4 114835.6 114833.2 114901.9 114915.6 110195.1 114963.5
#> [265] 114897.0 114314.5 109515.2 110064.9 114992.0 114836.0 114954.8 114789.7
#> [273] 114814.1 114913.3 114854.1 114980.8 114839.4 114865.8 114801.6 115128.3
#> [281] 113805.8 114801.4 114362.1 114978.9 114984.6 114823.8 114912.6 114668.8
#> [289] 115048.5 114869.5 115055.7 109153.7 114908.3 114765.1 114817.5 114862.8
#> [297] 114968.3 114985.2 113830.2 110110.2 110039.6 114930.3 114910.3 114829.7
#> [305] 110461.3 114785.6 114987.2 114829.3 114775.8 113643.1 110187.0 110213.0
#> [313] 115035.9 114837.7 114792.8 114892.1 114922.7 110057.1 114528.7 109361.8
#> [321] 109568.3 114760.7 114887.9 113947.6 109193.3 114466.2 114879.9 108793.4
#> [329] 114756.5 109048.6 115025.4 114775.7 114194.8 109237.4 114900.2 114212.2
#> [337] 114615.4 114206.4 114309.1 114246.8 114371.8 114295.2 114218.7 114327.6
#> [345] 114305.7 114156.5 114225.4 114267.9 114671.5 114220.1 114109.1 109091.9
#> [353] 114984.3 113948.5 114289.3 114948.4 111684.4 114937.8 114325.4 109183.1
#> [361] 115035.6 114897.8 114767.4 113602.4 110082.6 114306.7 114883.3 114992.8
#> [369] 114839.8 114702.0 114825.6 114925.2 114918.7 114840.8 114826.5 114866.8
#> [377] 115068.2 114893.5 114798.2 114891.2 109689.9 109453.3 114535.3 109568.6
#> [385] 114840.6 114014.0 114998.2 114337.3 108860.5 114303.1 114779.4 114345.2
#> [393] 114865.3 115047.7 114818.9 115015.7 114851.4 114971.6 110339.3 113745.5
#> [401] 108994.9 113792.8 114585.3 114899.7 115021.8 114952.1 109542.4 114802.2
#> [409] 113517.9 114918.0 114651.6 110419.5 114661.0 114737.6 109517.5 109432.3
#> [417] 113593.3 113529.0 114792.2 114754.3 114313.2 114923.7 114778.7 113608.5
#> [425] 113583.0 114877.1 114968.4 110231.2 114927.7 115121.9 114838.4 115077.9
#> [433] 114617.6 114780.0 114172.4 110175.6 114895.3 114884.5 114885.6 109111.7
#> [441] 114749.9 109697.2 114974.2 114731.2 114772.2 114835.4 114807.8 114803.4
#> [449] 114607.7 114329.9 114756.4 109557.7 114752.8 114684.9 109489.6 115030.1
#> [457] 114868.3 114787.6 114852.5 114850.0 114985.8 114946.1 114873.4 113890.6
#> [465] 114803.4 109589.6 114828.2 115027.4 113616.3 114951.2 114989.9 114919.6
#> [473] 108963.1 114751.2 113625.8 112829.5 114041.8 114806.0 114809.7 114801.0
#> [481] 114834.3 114726.9 113469.5 115007.1 114606.9 113385.9 114966.7 114765.4
#> [489] 110296.7 113908.4 109964.2 114783.7 114883.8 114867.3 112550.2 114980.9
#> [497] 113845.9 114737.8 115010.7 114816.5 114854.3 114821.3 109286.0 114828.1
#> [505] 114994.8 108886.9 108986.5 108937.5 114719.8 114301.0 114762.1 114809.2
#> [513] 114833.7 115088.8 114829.1 108950.0 114842.0 114749.1 114479.3 114837.2
#> [521] 115003.1 114526.0 109502.2 113550.9 114890.9 109517.1 109470.0 114803.0
#> [529] 113375.3 109485.0 109439.8 114876.2 109430.4 114607.2 109542.3 115036.5
#> [537] 109444.7 109505.0 114951.5 109537.9 109486.9 109497.5 109443.6 109563.5
#> [545] 109610.1 109080.5 114921.2 114869.1 114828.2 114901.7 110375.1 109470.8
#> [553] 114621.8 114616.3 114669.3 114576.0 114629.4 114728.1 109207.1 115029.9
#> [561] 114581.8 113854.5 114759.0 114594.5 114891.6 114115.8 114964.4 114864.6
#> [569] 113326.7 109510.1 114742.8 115008.8 114793.1 115052.3 114813.9 115019.9
#> [577] 114910.8 115060.4 114229.5 109529.6 114884.9 109154.0 114880.5 114805.8
#> [585] 114812.9 110282.6 114757.2 114977.4 114768.9 114825.9 114827.5 109432.1
#> [593] 109121.4 114667.3 110594.3 109283.0 115053.4 114670.4 114713.4 114211.2
#> [601] 114187.7 114889.1 114817.6 112966.7 113590.7 114722.7 114675.0 114606.4
#> [609] 113500.1 113872.6 111740.5 109863.0 114899.3 114947.1 115105.5 114765.7
#> [617] 110307.6 115068.3 108999.1 113370.7 113242.7 114896.9 114780.5 111728.4
#> [625] 114880.6 113356.5 114744.3 112560.6 115122.1 114769.9 114101.0 114068.0
#> [633] 114732.6 114941.5 115049.0 114883.0 109445.4 114912.9 114837.0 114867.9
#> [641] 109527.3 114189.4 114810.6 114845.0 114652.7 112687.8 114837.1 112961.5
#> [649] 114889.5 114964.9 115062.5 108941.0 114741.2 110189.3 112078.4 114872.8
#> [657] 114317.5 114966.7 114963.6 113969.2 115007.1 114985.7 114798.6 115029.2
#> [665] 114900.4 114943.4 114837.0 114741.2 114974.1 115037.6 108960.9 115079.4
#> [673] 114967.3 109502.6 114604.3 114891.5 114784.8 114892.9 114785.0 110210.8
#> [681] 114821.0 114655.1 115063.5 114710.9 110094.1 114955.7 115039.0 114894.6
#> [689] 115043.1 114978.3 114991.5 114878.3 114922.8 114742.3 114805.0 115018.5
#> [697] 114830.6 109549.6 114997.2 109416.0 114582.5 114960.9 114782.6 114862.9
#> [705] 114947.7 114820.9 114865.2 114797.2 114886.1 114873.5 109580.8 114622.6
#> [713] 111108.6 114885.6 111390.2 114827.2 114513.2 114885.7 115113.6 111700.8
#> [721] 114968.7 114625.5 114613.1 109762.1 110097.6 115103.9 114966.0 114786.2
#> [729] 114792.8 115019.7 114440.6 113936.8 114814.4 109470.2 114853.9 114770.5
#> [737] 114896.3 114982.5 114681.3 114942.7 114767.4 114664.0 112577.1 114943.1
#> [745] 113660.1 113811.8 114752.1 114841.7 114704.2 115077.8 114678.0 114812.3
#> [753] 110619.0 114705.3 109315.9 114971.0 114734.0 109507.7 114794.5 114736.6
#> [761] 114869.4 114899.7 114817.5 109220.9 109462.2 114952.0 114886.1 109127.3
#> [769] 114717.0 114835.4 114978.3 114931.1 114798.7 114858.2 114900.1 114751.8
#> [777] 114836.9 108881.0 111690.7 111729.7 114519.9 114735.2 114749.7 114215.7
#> [785] 109320.3 114757.7 114661.1 114807.6 115032.0 114668.7 114965.7 114176.8
#> [793] 114850.6 113584.3 110124.3 114946.2 114653.4 114996.3 114617.2 114743.4
#> [801] 114629.1 110117.9 114868.4 114817.8 114906.8 114880.6 114869.6 114778.4
#> [809] 114801.7 114756.1 109534.5 114696.2 115036.1 114761.2 114835.3 114846.4
#> [817] 114769.2 114817.4 114745.7 114736.5 114789.8 114790.9 109971.2 114810.8
#> [825] 109396.2 114901.7 114792.9 114747.7 113727.6 114828.0 114909.2 115012.3
#> [833] 115038.8 114874.5 109786.5 114819.4 114725.1 110136.9 114983.6 114883.0
#> [841] 114524.3 110926.2 114553.8 114958.5 114593.7 114643.5 114856.9 114823.3
#> [849] 114879.2 109353.8 109537.4 114756.6 115022.3 114851.1 110329.0 114900.0
#> [857] 113927.6 109483.7 114841.6 114730.0 115109.1 114904.5 109220.3 110268.4
#> [865] 114836.2 114823.2 113078.8 109336.4 114851.2 109317.7 109358.2 114793.8
#> [873] 114956.3 115073.5 115038.8 114768.9 114779.1 115091.9 114918.4 114826.1
#> [881] 109155.4 109082.1 114316.5 114918.6 110341.5 110948.2 114710.4 112771.3
#> [889] 109411.4 114789.9 109232.5 114762.0 114989.7 114731.3 114809.3 114831.6
#> [897] 114846.4 115040.2 113557.0 114766.7 113538.1 114800.8 114981.3 115091.2
#> [905] 114764.2 111166.7 110289.6 114134.2 114803.5 111812.2 114934.9 114962.9
#> [913] 108976.9 110857.6 114859.7 114731.1 114527.5 115042.6 109714.4 114831.9
#> [921] 115016.4 115001.4 114153.9 114902.0 109591.3 114060.0 114381.2 109160.0
#> [929] 114951.4 114842.8 114846.5 114794.0 114885.1 114785.5 114950.2 114672.4
#> [937] 114847.6 115012.1 114787.0 114214.4 114929.9 114550.3 114562.5 113993.1
#> [945] 115100.7 114255.6 114563.3 114542.8 115046.4 113128.9 114893.0 114811.3
#> [953] 114802.5 114737.6 114353.2 114789.7 114844.9 114847.1 114533.3 109469.2
#> [961] 109321.5 114928.6 114660.2 115020.8 114810.8 114799.6 109437.1 114518.3
#> [969] 114416.9 114670.2 114610.2 110057.9 114587.1 114832.9 114816.6 114188.6
#> [977] 115040.5 111646.1 114760.0 111481.3 114153.8 114232.1 114846.0 109253.2
#> [985] 111868.2 111377.7 115028.6 114887.1 114831.7 114620.7 114824.0 114988.0
#> [993] 114718.0 114653.0 109377.5 114815.3 114879.9 111171.8
#> 
#> 
#> [[3]]
#> [[3]]$residus
#> [1] 1739985952   80583045   29330311   31207521   39188586   47298094   41700301
#> [8]   55127023   36382145
#> 
#> [[3]]$delta.star.id
#> [1] 3
#> 
#> [[3]]$hat.X.delta.star
#>   [1] 5458.884 5399.712 5236.232 5720.513 5522.511 5511.104 5250.179 5188.762
#>   [9] 5137.936 5585.454 5294.343 5259.468 5735.746 5428.118 5243.605 5565.111
#>  [17] 5226.357 5688.770 5026.183 5090.711 5362.423 5225.091 5054.048 5094.657
#>  [25] 5478.371 5643.158 5227.795 5620.056 5372.565 5678.024 5125.764 5190.227
#>  [33] 5567.478 5018.782 5526.630 5005.055 5190.512 5690.240 5678.691 5186.171
#>  [41] 4747.425 5498.638 4980.375 5532.874 5232.476 5117.892 5101.017 5226.058
#>  [49] 5041.184 5434.841 5750.159 5572.162 5395.306 5265.742 5607.669 5637.467
#>  [57] 5091.280 5177.990 5650.964 4986.389 5454.545 5376.630 5281.813 5231.555
#>  [65] 5411.145 5654.382 5543.556 5154.760 5502.123 5402.679 5526.278 5491.549
#>  [73] 5260.665 5006.335 5661.574 5040.332 5507.845 5204.667 5270.384 5111.973
#>  [81] 4910.594 5060.615 5726.737 5562.339 5739.391 5517.497 5256.731 5023.141
#>  [89] 5504.854 5472.211 5209.640 5201.191 5490.206 5219.806 5559.075 5660.298
#>  [97] 5079.041 5841.553 5221.760 4965.895 5535.332 5016.194 5190.880 4989.522
#> [105] 5282.583 5676.008 4933.024 5252.049 5134.347 4788.719 5562.194 5762.329
#> [113] 5749.485 5019.468 4955.400 5012.366 5263.723 5292.325 5201.419 5191.644
#> [121] 5221.631 5450.951 5138.928 5748.972 4974.913 5184.601 5658.293 5019.457
#> [129] 5157.686 5693.465 5564.415 5644.029 5200.450 5051.891 5296.165 5202.848
#> [137] 5465.683 5204.504 5196.722 5677.632 5221.676 5428.008 5213.995 5096.919
#> [145] 5352.444 5424.292 5613.714 5496.115 5038.439 5682.476 5213.575 5203.521
#> [153] 5208.402 5216.773 5223.510 5427.709 5238.546 4985.483 5244.767 5219.752
#> [161] 5580.238 4975.754 5204.295 4896.803 4879.436 4896.620 5124.199 5554.991
#> [169] 5058.695 5237.617 5735.541 5047.489 5454.626 5542.182 5686.655 4970.142
#> [177] 5099.087 5597.562 5670.330 5571.871 5677.624 5231.393 5493.097 5527.423
#> [185] 5500.504 5425.070 5262.774 5643.333 5010.304 5571.015 5544.910 5205.099
#> [193] 5786.052 5017.436 4937.352 5474.207 5215.948 5181.133 5508.075 5570.497
#> [201] 5674.098 5784.699 5681.165 5213.890 5208.514 5296.182 5201.154 5482.216
#> [209] 5041.667 5604.028 4986.518 5640.835 5163.318 5197.122 5042.592 5418.928
#> [217] 5182.874 5680.831 5509.941 5760.332 5577.200 5277.627 5168.259 5538.736
#> [225] 5012.304 5216.325 5353.772 5541.195 5219.445 5286.557 5577.754 5033.877
#> [233] 5699.068 5310.778 5033.571 4994.924 5200.506 5205.059 5203.559 5015.097
#> [241] 5629.930 5188.768 5230.643 5844.022 5575.841 5700.814 5701.106 5451.194
#> [249] 5058.886 5196.676 5251.698 5204.289 5684.135 5261.756 5128.198 5459.968
#> [257] 5583.553 5220.715 4998.928 5036.303 5637.413 5421.260 5234.958 5672.590
#> [265] 5717.814 5233.320 5226.361 5208.222 5594.934 5273.627 5561.892 5155.877
#> [273] 5001.689 5555.552 5170.082 5662.305 5010.355 4990.926 5183.133 5745.471
#> [281] 5026.126 5112.973 5571.920 5647.201 5672.716 4980.482 5552.535 5132.303
#> [289] 5659.665 5531.574 5654.940 5231.719 5424.914 4984.014 5485.386 5542.462
#> [297] 5025.567 5651.431 5013.009 5206.230 5228.038 5564.419 5545.467 5362.962
#> [305] 5209.696 5030.257 5685.857 5173.316 5306.821 5012.246 5216.038 5230.215
#> [313] 5717.581 5310.685 5254.308 5613.518 5639.422 5214.537 4998.301 5212.320
#> [321] 5197.241 5288.432 5612.474 5091.262 5214.906 5161.823 5382.397 5271.518
#> [329] 5182.276 5282.579 5719.285 5026.328 5179.963 5237.431 5031.579 5167.904
#> [337] 5132.178 5161.822 5168.565 5171.313 5187.141 5161.350 5219.240 5217.655
#> [345] 5181.792 5208.372 5177.971 5152.851 5131.130 5191.056 5162.584 5264.123
#> [353] 5650.373 5110.113 5196.776 5677.741 5335.372 5775.956 5168.598 5231.473
#> [361] 5665.231 5541.565 5177.365 4944.020 5216.969 5212.972 5549.534 5821.881
#> [369] 5723.065 5198.196 5025.454 5518.893 5615.553 5559.668 5456.597 5463.643
#> [377] 5648.767 5476.524 5500.642 5490.161 5203.391 5194.920 5197.385 5189.163
#> [385] 5608.072 5102.207 5666.433 5110.087 5266.501 5226.431 5215.902 5238.940
#> [393] 5372.430 5716.713 5035.000 5664.010 5393.295 5662.472 5234.883 5093.638
#> [401] 5240.883 5039.629 4962.398 5623.334 5686.948 5518.611 5183.814 5031.921
#> [409] 4939.475 5450.979 5199.745 5178.836 4994.290 5161.124 5221.995 5201.424
#> [417] 4969.869 4951.681 5249.866 5046.946 5160.866 5459.178 5155.920 4965.823
#> [425] 4991.510 5461.183 5528.042 5181.349 5608.759 5817.886 4992.246 5748.676
#> [433] 5023.979 5298.429 5168.400 5203.741 5551.639 5499.715 5535.695 5213.298
#> [441] 5188.564 5166.854 5698.257 5200.568 5236.420 5519.207 5403.773 5371.573
#> [449] 5064.547 5184.620 5012.038 5205.731 5168.654 5234.703 5225.664 5689.034
#> [457] 5374.291 5037.755 5600.010 5468.850 5599.475 5602.020 5599.667 5071.256
#> [465] 5447.803 5219.596 5531.846 5650.179 4988.802 5596.187 5531.397 5490.455
#> [473] 5261.081 5173.810 4950.642 4762.935 5171.435 5027.313 5328.544 5240.819
#> [481] 5538.433 5218.598 4981.974 5609.937 5185.958 4895.532 5676.409 5315.038
#> [489] 5223.222 5060.900 5183.102 5036.808 5668.261 5561.780 5388.740 5593.736
#> [497] 5049.690 5015.654 5615.462 5459.835 5255.703 5323.255 5231.981 5011.866
#> [505] 5749.158 5258.086 5263.796 5186.342 4972.669 5238.719 5302.417 5432.275
#> [513] 5017.502 5725.164 5433.686 5243.658 5274.881 5279.896 5003.247 5205.226
#> [521] 5758.970 5202.549 5230.278 4950.429 5303.966 5214.360 5253.393 5655.014
#> [529] 4942.767 5263.772 5211.804 5381.058 5200.350 5185.719 5214.646 5640.896
#> [537] 5195.324 5180.359 5646.026 5237.009 5213.054 5220.388 5219.512 5182.741
#> [545] 5206.699 5258.926 5631.364 5425.034 5363.324 5508.795 5193.740 5212.594
#> [553] 5205.585 5190.519 5172.954 5170.345 5211.716 5158.291 5222.071 5666.507
#> [561] 5190.781 5065.931 5333.078 5058.182 5560.143 5207.439 5650.339 5629.732
#> [569] 4910.026 5229.231 5294.880 5787.563 5214.393 5703.845 5382.672 5664.060
#> [577] 5487.827 5797.025 5189.943 5210.083 5432.209 5260.092 5561.359 5410.834
#> [585] 5469.169 5208.305 4971.019 5397.272 5028.736 5368.693 5334.033 5237.556
#> [593] 5234.267 5195.437 5216.256 5231.036 5780.615 5198.511 5185.534 5146.335
#> [601] 5190.265 5276.794 5298.807 5459.023 4931.183 5145.573 5153.195 5055.224
#> [609] 4971.362 5119.251 5248.900 5192.609 5661.822 5477.328 5675.782 5170.240
#> [617] 5207.896 5693.742 5216.469 4944.041 4868.920 5013.996 5015.609 5237.478
#> [625] 4982.312 4905.616 5033.986 5374.616 5746.285 5287.774 5216.578 5228.654
#> [633] 5307.361 5602.652 5748.584 5628.812 5240.072 5458.116 5030.947 5285.369
#> [641] 5243.771 5219.219 5198.589 5456.391 5193.297 4733.056 5488.683 4761.247
#> [649] 5460.991 5705.889 5723.858 5254.366 5044.259 5217.164 5296.247 5627.000
#> [657] 5161.133 5575.886 5583.095 5050.967 5729.197 5591.306 5033.965 5623.472
#> [665] 5603.932 5628.254 5557.868 5009.280 5709.467 5663.677 5260.003 5684.478
#> [673] 5750.241 5189.939 5186.746 5544.246 5043.018 5087.587 5471.097 5202.575
#> [681] 5448.169 5061.569 5687.396 5459.457 5185.999 5598.763 5864.844 5388.423
#> [689] 5681.841 5676.003 5785.974 5499.356 5416.232 5000.766 4999.144 5671.133
#> [697] 5587.818 5205.388 5675.335 5231.729 5626.928 5656.572 5007.857 5499.700
#> [705] 5421.738 5059.806 5436.536 5314.363 5396.486 5471.115 5190.116 5147.995
#> [713] 5235.090 5165.912 5309.725 5362.357 5207.389 5507.171 5798.560 5270.030
#> [721] 5671.300 5009.214 5022.158 5182.514 5181.672 5764.037 5719.732 5020.884
#> [729] 5374.088 5651.211 5095.811 5059.266 4953.066 5256.824 5476.963 5228.891
#> [737] 5008.715 5477.762 5069.396 5456.227 5319.673 5155.723 4740.718 5717.792
#> [745] 5014.319 5023.713 5270.783 5159.965 4991.125 5706.385 5127.714 5329.556
#> [753] 5248.034 4985.451 5256.047 5494.015 5131.070 5217.857 5497.508 5350.114
#> [761] 5342.057 5525.242 5233.860 5230.019 5243.470 5559.650 5469.410 5239.391
#> [769] 5151.108 5237.436 5724.387 5510.840 5250.878 5233.466 5233.836 5276.782
#> [777] 5323.196 5235.854 5299.093 5280.526 5221.707 5210.333 5545.456 5231.860
#> [785] 5231.044 5034.616 5128.057 5359.425 5684.430 4989.700 5542.805 5173.764
#> [793] 5357.060 5002.616 5212.316 5509.693 5192.294 5689.872 5211.852 5014.847
#> [801] 5185.159 5250.013 5335.175 5465.705 5344.920 5565.561 5469.347 4983.952
#> [809] 5231.510 5293.763 5197.631 5223.209 5717.585 5149.741 5547.722 5450.445
#> [817] 5228.328 5143.455 4968.634 5283.403 5126.189 5411.257 5223.547 5381.477
#> [825] 5245.148 5557.845 5559.483 5320.338 5048.988 5561.388 5630.110 5685.563
#> [833] 5724.188 5517.593 5171.657 5016.533 5186.891 5202.018 5717.820 5548.670
#> [841] 5178.054 5257.601 5026.341 5672.432 4992.386 4995.286 5134.412 5278.890
#> [849] 5495.624 5230.860 5206.085 5351.346 5668.731 5454.730 5229.411 5559.954
#> [857] 5094.561 5244.234 5332.482 5282.970 5674.214 5408.381 5252.353 5228.809
#> [865] 5129.581 5495.659 5482.600 5249.577 5393.432 5246.035 5253.591 5149.326
#> [873] 5637.341 5729.061 5725.935 5223.243 5114.178 5691.112 5599.112 5327.423
#> [881] 5218.777 5237.737 5164.848 5024.719 5236.549 5250.604 5165.877 4745.028
#> [889] 5240.901 5157.666 5188.256 5281.947 5655.461 5011.714 5133.213 5158.365
#> [897] 5553.939 5640.190 4967.900 5348.760 4941.764 5521.031 5667.599 5675.283
#> [905] 4984.962 5266.993 5207.280 5201.084 5027.364 5305.698 5505.719 5512.592
#> [913] 5269.390 5257.585 5494.719 5197.411 5217.582 5681.224 5194.027 5341.846
#> [921] 5697.789 5675.602 5171.290 5345.100 5241.417 5143.269 5207.062 5212.946
#> [929] 5610.867 5044.158 5047.777 5160.768 5291.803 5322.339 5717.660 5198.016
#> [937] 5407.551 5725.873 5368.578 5158.798 5631.990 5202.447 5188.087 5096.521
#> [945] 5750.926 5208.698 5162.307 5223.995 5707.087 4812.064 5610.749 5040.456
#> [953] 5026.741 5010.072 5252.270 5206.622 5373.559 5282.883 5216.752 5217.838
#> [961] 5258.079 5539.485 5253.740 5698.195 5001.289 5265.318 5199.715 5220.733
#> [969] 5194.080 5216.127 5218.252 5204.932 5174.654 5561.589 5393.487 5269.560
#> [977] 5711.331 5274.837 5412.880 5278.865 5253.442 5229.412 5428.988 5248.858
#> [985] 5268.756 5266.837 5677.414 5607.730 5413.261 5658.837 5538.460 5648.595
#> [993] 5335.280 4989.790 5247.301 5423.094 5409.992 5273.763
#> 
#>